% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{alluvial.plot}
\alias{alluvial.plot}
\title{Alluvial plot}
\usage{
alluvial.plot(
  bsrinf,
  keywords = c("COL4A1"),
  type = c("L", "R", "pw.id"),
  qval = 1,
  format = c("svg", "png", "pdf"),
  path = "./",
  filename = "alluvial.plot",
  width = 10,
  height = 10
)
}
\arguments{
\item{keywords}{vector of pathways.}

\item{type}{filter on Ligand, Recptor or pathway id.}

\item{qval}{threshold over Q-value.}

\item{format}{svg / png / pdf. By default means it will 
plot in a svg format.}

\item{path}{directory where to plot file.}

\item{filename}{file name.}

\item{width}{width of image.}

\item{height}{height of image.}

\item{bsrsig}{object bsrinf inference.}
}
\value{
NULL (plot printed somewhere on disk)
This is a convenience function that relies on the \code{ggalluvial}
package to propose a simple way
of representing Ligands, Receptors
}
\description{
Representation of the links
between Ligands,Receptors and Pathways.
}
\examples{
print('alluvial.plot')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
alluvial.plot(bsrinf,
             keywords = c("COL4A1"),
             type = "L",
             qval = 0.001,
             path = "./",
             filename = "sdc_alluvial", 
             width  = 16, 
             height = 12
             )
}
