% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotation.spa}
\alias{annotation.spa}
\title{A skinny data frame used in the spatial workflow}
\format{
Data frame that contains the following columns:
barcode_id,sample_id, in_tissue,array_row
array_col, ground_truth, reference, cell_count, idSpatial

barcode_id is the id of the spot
idSpatial is the spatial id of the spot(array_rowXarray_col)
ground_truth is the label (Layer1/2 were only kept)

They are the mandatory data necessary to
generate plots for the spatial workflow.
}
\source{
\url{http://spatial.libd.org/spatialLIBD/}
}
\usage{
data(annotation.spa)
}
\description{
Data frame subset describing the spatial spots
}
\keyword{datasets}
