% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModelComp-class.R
\name{as.BSRDataModelComp}
\alias{as.BSRDataModelComp}
\title{conversion of BSRDataModel into BSRDataModelComp}
\usage{
as.BSRDataModelComp(bsrdm)
}
\arguments{
\item{bsrdm}{A BSRDataModel object.}
}
\description{
In case ligand-receptor inferences should be obtained
based on gene/protein regulation P-values comparing two clusters of
samples, it is necessary to first promote the BSRDataModel object that
contains the count matrix into a BSRDataModelComp object able to contain
a list of such cluster pairs comparisons. This function performs this
promotion adding an empty list of comparisons.
}
\examples{
# prepare data
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)

}
