% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\docType{class}
\name{BSRClusterComp-class}
\alias{BSRClusterComp-class}
\title{BulkSignalR Cluster Comparison Object}
\description{
An S4 class to represent the comparison of two clusters of samples to
infer LR interactions based on the resulting P-values,
log-fold-changes (logFC), and expression values.
}
\section{Slots}{

\describe{
\item{\code{col.clusterA}}{Column indices for the samples in cluster A.}

\item{\code{col.clusterB}}{Column indices for the samples in cluster B.}

\item{\code{differential.stats}}{Comparison statistics A versus B 
as a data.frame and
containing at least 3 columns named 'pval', 'logFC', and 'expr'.}
}}

\examples{
new("BSRClusterComp")
}
