% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModelComp-class.R
\docType{class}
\name{BSRDataModelComp-class}
\alias{BSRDataModelComp-class}
\title{BulkSignalR Data Model Compare Object}
\description{
An S4 class to represent the expression data used for inferring
ligand-receptor interactions based on sample cluster comparisons.
}
\section{Slots}{

\describe{
\item{\code{comp}}{A named list of BSRClusterComp objects, one per
comparison.}

\item{\code{mu}}{A number representing the average value in the normalized and
lop1p-transformed gene expression matrix. This value is used to compute
the LR-score (cf. SingleCellSignalR paper, Cabello-Aguilar, et al.,
Nucleic Acids Res, 2020)}
}}

\examples{
new("BSRDataModelComp")

}
