% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsrdm.comp}
\alias{bsrdm.comp}
\title{A skinny BSRDataModelComp object related to sdc.}
\format{
An example of an BSRDataModelComp object
}
\usage{
data(bsrdm.comp)
}
\description{
See Vignette BulkSignalR-Differential.
}
\keyword{datasets}
