% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsrdm}
\alias{bsrdm}
\title{A skinny BSRDataModel object related to sdc.}
\format{
An example of an object created by `BSRDataModel`
applied to an sdc subset (Patients N20,N22,SDC17,SDC25) and
10 000 genes sampled (seed set to 123)
`learnParameters` was also called to get statistical 
model parameters.
}
\usage{
data(bsrdm)
}
\description{
Output from the `learnParameters` function to get
BulkSignalR statistical model parameters.
}
\keyword{datasets}
