% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsrinf.example}
\alias{bsrinf.example}
\title{A BSR-Inference object related to sdc.}
\format{
An example of an object created by inference function
}
\usage{
data(bsrinf.example)
}
\description{
From the previous object `bsrdm`, 
you can generate inferences by calling its
method `initialInference`. 
The resulting BSR-Inference object is `bsrinf`,
It contains all the inferred L-R interactions 
with their associated pathways and corrected p-values.
}
\keyword{datasets}
