% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsrinf.mouse}
\alias{bsrinf.mouse}
\title{A skinny BSRInference object related to bodyMap.mouse}
\format{
An example of an object created by inference function
}
\usage{
data(bsrinf.mouse)
}
\description{
see related workflow for non human organism
in the vignette
}
\keyword{datasets}
