% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsrinf}
\alias{bsrinf}
\title{A skinny BSRInference object related to sdc.}
\format{
An example of an object created by inference function
}
\usage{
data(bsrinf)
}
\description{
From the previous object `bsrdm`, 
you can generate inferences by calling its
constructor `BSRInference`. 
The resulting BSRInference object is `bsrinf`,
It contains all the inferred L-R interactions 
with their associated pathways and corrected p-values.
}
\keyword{datasets}
