% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsrinf.spa}
\alias{bsrinf.spa}
\title{A skinny BSR-inference object related to spatial data set}
\format{
From the previous object `bsrdm.spa`, 
you can generate inferences by calling its
method `BSRInference`. 
The resulting BSRInference object is `bsrinf.spa`,
It contains all the inferred L-R interactions 
with their associated pathways and corrected p-values.
`learnParameters` was also called to train the statistical 
model parameters.
}
\source{
\url{http://spatial.libd.org/spatialLIBD/}
}
\usage{
data(bsrinf.spa)
}
\description{
Output from the `learnParameters` function to get
BulkSignalR statistical model parameters.
}
\keyword{datasets}
