% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRSignature-class.R
\docType{class}
\name{BSRSignature-class}
\alias{BSRSignature-class}
\title{BulkSignalR ligand-receptor signature Object}
\description{
S4 class to represent gene signatures
of inferred ligand-receptor interactions, including
their reduced versions.
}
\section{Slots}{

\describe{
\item{\code{ligands}}{A list of ligands, one entry per LR interaction.}

\item{\code{receptors}}{A list of receptors, one entry per LR interaction.}

\item{\code{t.genes}}{A list of target genes, one entry per LR interaction.}

\item{\code{pathways}}{An atomic vector of pathway names, one per interaction.}

\item{\code{tg.corr}}{A list of target genes correlation.}
}}

\examples{
new("BSRSignature")

}
