% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRSignatureComp-class.R
\docType{class}
\name{BSRSignatureComp-class}
\alias{BSRSignatureComp-class}
\title{BulkSignalR ligand-receptor signature object for cluster comparisons}
\description{
S4 class to represent gene signatures associated with ligand-receptor
interactions that were inferred from the comparison of two clusters
of samples. This class inherits from BSRSignature.
}
\section{Slots}{

\describe{
\item{\code{cmp.name}}{The name of the comparison.}

\item{\code{tg.pval}}{A list of target genes P-values.}

\item{\code{tg.logFC}}{A list of target genes logFC.}
}}

\examples{
new("BSRSignatureComp")

}
