% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{bubblePlot.pathways.LR}
\alias{bubblePlot.pathways.LR}
\title{Bubble Plot to explore LR & Pathways}
\usage{
bubblePlot.pathways.LR(
  bsrinf,
  pathways = c("Cell surface interactions at the vascular wall"),
  threshold = 1,
  filter.L = NULL,
  filter.R = NULL,
  path = "./",
  filename = "DotPlot",
  color = "#16a647",
  width = 16,
  height = 7,
  pointsize = 6,
  format = c("svg", "png", "pdf")
)
}
\arguments{
\item{bsrinf}{BulkSignalR inference object.}

\item{pathways}{Vector or pathway names}

\item{threshold}{BulkSignalR data model object.}

\item{filter.L}{Filter a list of ligands.}

\item{filter.R}{Filter a list of receptors.}

\item{path}{Path directory to plot.}

\item{filename}{An output filename.}

\item{color}{Main color used for the gradient.}

\item{width}{Global image width size.}

\item{height}{Global image height size.}

\item{pointsize}{Global pointsize.}

\item{format}{File format svg (default),png or pdf.}
}
\value{
A plot is created with the given filename. 

This is a convenience function to propose a simple way
of representing LR - Pathways association
with their respective correlation
and Q-values.
}
\description{
Quick check to observe LR - Pathways association
with their respective correlation
and Q-values.
}
\examples{
print('bubblePlot.pathways.LR')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
pathways <- c("PD-1 signaling","Interferon gamma signaling")
bubblePlot.pathways.LR(bsrinf,
   pathways = pathways, 
   threshold = 1,
   path = "./",
   color = "red",
   filename  = "sdc_bubble", 
   width  = 16, 
   height = 7,
   pointsize = 8
   )  
}
