% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkSignalR.R
\docType{package}
\name{BulkSignalR-package}
\alias{BulkSignalR}
\alias{BulkSignalR-package}
\title{BulkSignalR: A package to infer ligand-receptor
interactions from bulk transcriptomics or proteomics}
\description{
Inference of ligand-receptor interactions from bulk
(transcriptomic or proteomic) data. BulkSignalR bases its inferences
on the LRdb database. It relies on a statistical model that
is specific to bulk data sets. Different visualization and data
summary functions are proposed to help navigating results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ZheFrench/BulkSignalR}
  \item Report bugs at \url{https://github.com/ZheFrench/BulkSignalR/issues}
}

}
\author{
\strong{Maintainer}: Jean-Philippe Villemin \email{jpvillemin@gmail.com} (\href{https://orcid.org/0000-0002-1838-5880}{ORCID})

Authors:
\itemize{
  \item Jacques Colinge \email{jacques.colinge@inserm.fr} (\href{https://orcid.org/0000-0003-2466-4824}{ORCID})
}

}
\keyword{internal}
