% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{checkReceptorSignaling}
\alias{checkReceptorSignaling}
\title{Check receptor signaling downstream}
\usage{
checkReceptorSignaling(
  ds,
  lr,
  method = c("reactome-GOBP", "reactome", "GOBP")[1],
  max.pw.size = 200,
  min.pw.size = 5,
  min.positive = 4,
  restrict.pw = NULL,
  with.complex = TRUE
)
}
\arguments{
\item{ds}{A BulkSignalR data set.}

\item{lr}{A ligand-receptor analysis such as returned by \code{getCorrelatedLR()}.}

\item{method}{Which pathway reference should be used ("reactome" for Reactome,"GOBP" for GO Biological Process,
or "Reactome-GOBP" for both).}

\item{max.pw.size}{Maximum pathway size to consider from the pathway reference.}

\item{min.pw.size}{Minimum pathway size to consider from the pathway reference.}

\item{min.positive}{Minimum number of target genes to be found in a given pathway.}

\item{restrict.pw}{A list of pathway IDs to restrict the application of the function.}

\item{with.complex}{A logical indicating whether receptor co-complex members should be included in the target genes.}
}
\value{
A ligand-receptor analysis extending \code{lr} content with a table reporting all the ligand-receptor pairs provided in \code{lr}
along with the pathways found and data about target gene correlations with the receptor. Those data are the lists
(in semi-colon-separated format) of target genes and their Spearman correlations with the receptor. The lists are sorted
according to the correlation coefficient.

In a pathway of the reference, i.e., a Reactome pathway or the genes of a GOBP term, the target genes are the
genes coding for proteins forming a complex with the receptor and the genes in the pathway downstream the receptor,
which are given as regulated by the pathway. If \code{with.complex} is set to \code{FALSE}, then only the
regulated genes are considered. Participation to a complex and being regulated as well as the pathway directed topologies
are defined by Reactome and KEGG pathways as provided by PathwayCommons.

The maximum pathway size is used to limit the redundancy inherent to GOBP and Reactome. The minimum pathway size is
used to avoid overspecific, noninformative results.
}
\description{
Assess the existence of correlations between a receptor, part of a ligand-receptor pair, and
genes coding for proteins forming a complex with the receptor or genes regulated by the
receptor downstream signaling.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
}
}
