% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R
\name{cmpName<-,BSRInferenceComp-method}
\alias{cmpName<-,BSRInferenceComp-method}
\title{Comparison name setter (internal use only)}
\usage{
\S4method{cmpName}{BSRInferenceComp}(x) <- value
}
\arguments{
\item{x}{BSRInferenceComp object}

\item{value}{value to be set for bsrinf}
}
\description{
Comparison name setter (internal use only)
}
\keyword{internal}
