% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModelComp-class.R
\name{coerce}
\alias{coerce}
\alias{coerce,BSRDataModel,BSRDataModelComp-method}
\title{Convert BSRDataModel to BSRDataModelComp}
\arguments{
\item{from}{BSRDataModel object}
}
\value{
A BSRDataModelComp object
}
\description{
Enable the promotion of a BSRDataModel object into
a BSRDataModelComp object by
coercion.
}
\examples{
bsrdm <- new("BSRDataModel")
bsrdm.comp <- as(bsrdm, "BSRDataModelComp")

}
