% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{colA<-,BSRClusterComp-method}
\alias{colA<-,BSRClusterComp-method}
\title{Cluster A columns setter (internal use only)}
\usage{
\S4method{colA}{BSRClusterComp}(x) <- value
}
\arguments{
\item{x}{object BSRClusterComp}

\item{value}{value to be set for BSRClusterComp}
}
\description{
Cluster A columns setter (internal use only)
}
\keyword{internal}
