% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{colA}
\alias{colA}
\alias{colA,BSRClusterComp-method}
\title{Cluster A columns accessor}
\usage{
\S4method{colA}{BSRClusterComp}(x)
}
\arguments{
\item{x}{object BSRClusterComp}
}
\description{
Cluster A columns accessor
}
