% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{colB<-,BSRClusterComp-method}
\alias{colB<-,BSRClusterComp-method}
\title{Cluster B columns setter (internal use only)}
\usage{
\S4method{colB}{BSRClusterComp}(x) <- value
}
\arguments{
\item{x}{object BSRClusterComp}

\item{value}{value to be set for BSRClusterComp}
}
\description{
Cluster B columns setter (internal use only)
}
\keyword{internal}
