% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{colB}
\alias{colB}
\alias{colB,BSRClusterComp-method}
\title{Cluster B columns accessor}
\usage{
\S4method{colB}{BSRClusterComp}(x)
}
\arguments{
\item{x}{object BSRClusterComp}
}
\value{
colB
}
\description{
Cluster B columns accessor
}
\examples{
bsrdm <- new("BSRDataModel",
    ncounts = matrix(1.5,
        nrow = 2, ncol = 4,
        dimnames = list(c("A", "B"), c("E", "F","G","H"))
    ),
    log.transformed = TRUE, normalization = "TC"
)
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:2)
colB <- as.integer(3:4)
n <- nrow(ncounts(bsrdm.comp))
edger.stats <- data.frame(
    pval = runif(n), logFC = rnorm(n, 0, 2),
    expr = c(1, 2)
)
rownames(edger.stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, edger.stats)
colB(bsrcc)
}
