% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{colClusterA}
\alias{colClusterA}
\alias{colClusterA,BSRClusterComp-method}
\title{Cluster A columns accessor}
\usage{
\S4method{colClusterA}{BSRClusterComp}(x)
}
\arguments{
\item{x}{object BSRClusterComp}
}
\value{
col.clusterA
}
\description{
Cluster A columns accessor
}
\examples{
bsrcc <- new("BSRClusterComp")
colClusterA(bsrcc)
}
