% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{colClusterB<-,BSRClusterComp-method}
\alias{colClusterB<-,BSRClusterComp-method}
\title{Cluster B columns setter (internal use only)}
\usage{
\S4method{colClusterB}{BSRClusterComp}(x) <- value
}
\arguments{
\item{x}{object BSRClusterComp}

\item{value}{value to be set for BSRClusterComp}
}
\value{
returns \code{NULL}
}
\description{
Cluster B columns setter (internal use only)
}
\keyword{internal}
