% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{colClusterB}
\alias{colClusterB}
\alias{colClusterB,BSRClusterComp-method}
\title{Cluster B columns accessor}
\usage{
\S4method{colClusterB}{BSRClusterComp}(x)
}
\arguments{
\item{x}{object BSRClusterComp}
}
\value{
col.clusterB
}
\description{
Cluster B columns accessor
}
\examples{
bsrcc <- new("BSRClusterComp")
colClusterB(bsrcc)
}
