% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModelComp-class.R
\name{comparison<-,BSRDataModelComp-method}
\alias{comparison<-,BSRDataModelComp-method}
\title{Comparisons list setter (internal use only, use addComparison() otherwise)}
\usage{
\S4method{comparison}{BSRDataModelComp}(x) <- value
}
\arguments{
\item{x}{object BSRDataModelComp}

\item{value}{value to be set for BSRDataModelComp}
}
\value{
returns \code{NULL}
}
\description{
Comparisons list setter (internal use only, use addComparison() otherwise)
}
\keyword{internal}
