% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModelComp-class.R
\name{comparison}
\alias{comparison}
\alias{comparison,BSRDataModelComp-method}
\title{Comparisons list accessor}
\usage{
\S4method{comparison}{BSRDataModelComp}(x)
}
\arguments{
\item{x}{object BSRDataModelComp}
}
\value{
comp
}
\description{
Comparisons list accessor
}
\examples{
bsrdm.comp <- new("BSRDataModelComp")
comparison(bsrdm.comp)
}
