% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrepare.R
\name{convertToHuman}
\alias{convertToHuman}
\title{Transpose to Human Gene Names}
\usage{
convertToHuman(
  counts,
  dictionary = data.frame(Gene.name = "A", row.names = "B")
)
}
\arguments{
\item{counts}{A table or matrix of read counts.}

\item{dictionary}{A dataframe where first column belong to 
organism of study & rownames are the human gene names.}
}
\value{
Return a counts matrix transposed for Human.
}
\description{
By default, BulkSignalR is designed to work with Homo Sapiens.
In order to work with other species, gene names need to be first converted
to Human following an orthology mapping process.
}
\examples{
print('convertToHuman')
data(bodyMap.mouse)

ortholog.dict    <- findOrthoGenes (from_organism = "mmusculus", 
                                    from_values = rownames(bodyMap.mouse))

matrix.expression.human <- convertToHuman(counts = bodyMap.mouse,   
dictionary = ortholog.dict)

}
