% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{coordsFlip}
\alias{coordsFlip}
\title{(counter-)clockwise 90 degree rotation}
\usage{
coordsFlip(rasterImage, degrees = 90)
}
\arguments{
\item{rasterImage}{raster image}

\item{degrees}{degrees of rotation}
}
\value{
A rotated raster image
}
\description{
(counter-)clockwise 90 degree rotation
}
\examples{
print('coordsFlip')
if(FALSE){
img <- png::readPNG(path.to.file)
my.image.as.raster <- grDevices::as.raster(path.to.file)  
my.image.rotated <- coordsFlip(my.image.as.raster,degrees=90)

 }
}
