% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{createDatabase}
\alias{createDatabase}
\title{Fetch the LR reference database from internet}
\usage{
createDatabase(onRequest = TRUE, verbose = FALSE)
}
\arguments{
\item{onRequest}{logical TRUE to force
downloading again. This will overwrite the
pre-existing local database. Default is TRUE.}

\item{verbose}{Logical}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Fetch LR database from remote location.
}
\examples{
print("Function already called elsewhere by cacheClear()")
# createDatabase(onRequest = FALSE)
}
