% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModelComp-class.R
\name{defineClusterComp}
\alias{defineClusterComp}
\alias{defineClusterComp,BSRDataModelComp-method}
\title{Definition of the comparison between two clusters of samples}
\usage{
\S4method{defineClusterComp}{BSRDataModelComp}(obj, colA, colB, stats)
}
\arguments{
\item{obj}{A BSRDataModelComp object output by
\code{\link{as.BSRDataModelComp}}.}

\item{colA}{Cluster A column indices.}

\item{colB}{Cluster B column indices.}

\item{stats}{A data.frame containing statistics about the differential
analysis cluster A versus B. \code{stats} must contain at least the
columns 'pval' (for P-values) and 'logFC' for log-fold-changes A/B.}
}
\value{
A BSRClusterComp object.
}
\description{
Define the columns of the expression matrix that belong to each cluster,
and store the result of the cluster differences statistical analysis
obtained by an external tool such as edgeR, DESeq2, etc.
}
\details{
Create a BSRClusterComp object describing a comparison
of two clusters of columns taken from the expression matrix
in the BSRDataModelComp object \code{obj}. Such a cluster comparison
description is the basis for inferring LRIs from differential
expression P-values instead of correlation analysis.

The rows of \code{stats} must be in the same order as those of the count
matrix in \code{obj}. Alternatively, \code{stats}
rows can be named and a 1-1 correspondence must exist between these names
and those of the count matrix.
}
\examples{
# prepare data
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:5)
colB <- as.integer(8:15)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(pval=runif(n), logFC=rnorm(n, 0, 2))
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison
bsrinf <- initialInference(bsrdm.comp, max.pval=1,"random.example")

}
