% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{differentialStats<-,BSRClusterComp-method}
\alias{differentialStats<-,BSRClusterComp-method}
\title{Cluster comparison statistics setter (internal use only)}
\usage{
\S4method{differentialStats}{BSRClusterComp}(x) <- value
}
\arguments{
\item{x}{object BSRClusterComp}

\item{value}{value to be set for BSRClusterComp}
}
\value{
returns \code{NULL}
}
\description{
Cluster comparison statistics setter (internal use only)
}
\keyword{internal}
