% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{differentialStats}
\alias{differentialStats}
\alias{differentialStats,BSRClusterComp-method}
\title{Cluster comparison statistics accessor}
\usage{
\S4method{differentialStats}{BSRClusterComp}(x)
}
\arguments{
\item{x}{BSRClusterComp object}
}
\value{
diffferential.stats
}
\description{
Cluster comparison statistics accessor
}
\examples{
bsrcc <- new("BSRClusterComp")
differentialStats(bsrcc)
}
