% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cacheCheckIn}
\alias{.cacheCheckIn}
\title{Check existence of a record in the cache.}
\usage{
.cacheCheckIn(bfc, resourceName)
}
\arguments{
\item{bfc}{Object of class BiocFileCache, created by a call to
BiocFileCache::BiocFileCache()}

\item{resourceName}{keyword associated to a specific resource name.}
}
\value{
logical This function returns TRUE if a record with
the requested keyword already exists in the file cache,
 otherwise it returns FALSE.
}
\description{
Check whether the cache record exists or not by passing
to the function an associated keyword
related to the resource we are looking for.
}
\keyword{internal}
