% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.cdfAlphaStable}
\alias{.cdfAlphaStable}
\title{Internal function to compute a censored alpha-) stable CDF}
\usage{
.cdfAlphaStable(x, par)
}
\arguments{
\item{x}{A vector of observed values.}

\item{par}{A list containing the censored stable model parameters.}
}
\value{
A vector of probabilities P(X<x|par).
}
\description{
Internal function to compute a censored alpha-) stable CDF
}
\keyword{internal}
