% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.cdfEmpirical}
\alias{.cdfEmpirical}
\title{Internal function to compute an empirical CDF}
\usage{
.cdfEmpirical(x, par)
}
\arguments{
\item{x}{A vector of observed values.}

\item{par}{A list containing the step function implementing the CDF.}
}
\value{
A vector of probabilities P(X<x|par).
}
\description{
Internal function to compute an empirical CDF
}
\keyword{internal}
