% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.cdfGaussian}
\alias{.cdfGaussian}
\title{Internal function to compute a censored Gaussian CDF}
\usage{
.cdfGaussian(x, par)
}
\arguments{
\item{x}{A vector of observed values.}

\item{par}{A list containing the censored Gaussian model parameters.}
}
\value{
A vector of probabilities P(X<x|par).
}
\description{
Internal function to compute a censored Gaussian CDF
}
\keyword{internal}
