% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.cdfKernelEmpirical}
\alias{.cdfKernelEmpirical}
\title{Internal function to compute a Gaussian kernel-based empirical CDF}
\usage{
.cdfKernelEmpirical(x, par)
}
\arguments{
\item{x}{A vector of observed values.}

\item{par}{A list containing the step function implementing the CDF.}
}
\value{
A vector of probabilities P(X<x|par).
}
\description{
Internal function to compute a Gaussian kernel-based empirical CDF
}
\keyword{internal}
