% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.cdfMixedGaussian}
\alias{.cdfMixedGaussian}
\title{Internal function to compute a censored mixed-Gaussian CDF}
\usage{
.cdfMixedGaussian(x, par)
}
\arguments{
\item{x}{A vector of observed values.}

\item{par}{A list containing the censored mixed-Gaussian model parameters.}
}
\value{
A vector of probabilities P(X<x|par).
}
\description{
Internal function to compute a censored mixed-Gaussian CDF
}
\keyword{internal}
