% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{.checkDatabaseValidity}
\alias{.checkDatabaseValidity}
\title{Check validity of database}
\usage{
.checkDatabaseValidity(connexionObject)
}
\arguments{
\item{connexionObject}{DBI::dbCanConnect object
to test if it's a valid connexion.}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Control connexion is ok.
}
\keyword{internal}
