% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.checkRDSFromCache}
\alias{.checkRDSFromCache}
\title{Check for valid RDS cache file.}
\usage{
.checkRDSFromCache(bfc, resourceName)
}
\arguments{
\item{bfc}{Object of class BiocFileCache, created by a call to
BiocFileCache::BiocFileCache()}

\item{resourceName}{keyword associated to a specific resource name.}
}
\value{
logical TRUE/FALSE
}
\description{
This function checks whether a cache entry is a valid RDS file.
Returns TRUE if the cache entry is valid, FALSE otherwise.
In the case of an invalid file, the cache entry and file are
deleted.
}
\keyword{internal}
