% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.cutExtremeValues}
\alias{.cutExtremeValues}
\title{Internal function to cut extreme values from a matrix}
\usage{
.cutExtremeValues(m, p)
}
\arguments{
\item{m}{A matrix.}

\item{p}{Proportion of top and bottom values for thresholding.}
}
\value{
A matrix with values beyond top and bottom thresholds repaced by
the latter thresholds.
}
\description{
Internal function to cut extreme values from a matrix
}
\keyword{internal}
