% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{.downloadDatabase}
\alias{.downloadDatabase}
\title{Fetch the database from internet.}
\usage{
.downloadDatabase(url, databaseFilePath)
}
\arguments{
\item{url}{File URL.}

\item{databaseFilePath}{Path to database file.}
}
\description{
Fetch LR database from remote location.
}
\keyword{internal}
