% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.formatPathwaysFromGmt}
\alias{.formatPathwaysFromGmt}
\title{Transform gmt file to data frame}
\usage{
.formatPathwaysFromGmt(file, resourceName = NULL)
}
\arguments{
\item{file}{Path to GMT file}

\item{resourceName}{Two options "GO-BP" or "REACTOME"}
}
\value{
Dataframe with pathwayID, geneName and pathwayName
}
\description{
We note discrepancies between the formats available
from internet sources. Here, we consider a valid gmt file format defined
on each lines as follows:
First is Pathway name,
then comes the ID,
finally you will find genes symbols
part of the pathway defined on the line.
}
\details{
You can find an example here.
- For Reactome. (Directly from their website)
\url{https://reactome.org/download/current/ReactomePathways.gmt.zip}
Note that you need to unzip the file to read the content.
The code is inspired from read.gmt function
from the gsa R package.
}
