% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.formatPathwaysFromJson}
\alias{.formatPathwaysFromJson}
\title{Format a data frame according to json input}
\usage{
.formatPathwaysFromJson(file, resourceName = NULL)
}
\arguments{
\item{file}{Path to file.}

\item{resourceName}{Two options "GO-BP" or "REACTOME".}
}
\value{
Data frame with pathwayID, geneName and pathwayName
}
\description{
Format a data frame according to json input
}
