% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{.geneNameConversion}
\alias{.geneNameConversion}
\title{Convert gene symbols to another organism}
\usage{
.geneNameConversion(genes, conversion.dict)
}
\arguments{
\item{genes}{The genes you want to convert.}

\item{conversion.dict}{A data frame containing
gene names for the source species and Homo sapiens.}
}
\value{
Depend on the type of input genes
LRinter return a vector of genes
tgGenes receptors ligands: return list of list of genes
}
\description{
Convert gene symbols to another organism
based on a dictionary with human and orthologs in the other species.
}
\keyword{internal}
