% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.getDifferentialPathways}
\alias{.getDifferentialPathways}
\title{Internal function to search for differentially regulated pathways}
\usage{
.getDifferentialPathways(
  ncounts,
  types,
  normal,
  induced,
  max.pw.size = 200,
  min.pw.size = 5,
  max.qval = 0.01,
  min.FC = log2(1.5),
  min.count = 100
)
}
\arguments{
\item{ncounts}{A matrix or table of normalized read counts.}

\item{types}{A character vector defining the sample types.}

\item{normal}{A character vector defining the sample types that correspond to the "normal" condition}

\item{induced}{A character vector defining the sample types that correspond to the "induced" condition}

\item{max.pw.size}{Maximum pathway size to consider from the pathway reference.}

\item{min.pw.size}{Minimum pathway size to consider from the pathway reference.}

\item{max.qval}{Maximum Q-value imposed to genes.}

\item{min.FC}{Minimum log2 fold-change (in absolute value) imposed to genes.}

\item{min.count}{Minimum median expression value across all the samples imposed to genes.}
}
\value{
A table of pathways as returned by \code{\link{.pwHyperScores}}.

The 2-condition induced versus normal is performed using edgeR \code{\link[edgeR]{glmFit}}.

No normalization is performed with edgeR since the matrix is supposed to be normalized already. Note that it is possible
to provide \code{\link{prepareDataset}} with a matrix that would have been pre-normalized using
\code{\link[edgeR]{calcNormFactors}} if so desired. (Remember that \code{\link[edgeR]{calcNormFactors}} does not normalize the matrix, it only
returns the normalization factors.)
}
\description{
Perform differential gene analysis with edgeR and score pathways.
}
