% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.getEmpiricalParam}
\alias{.getEmpiricalParam}
\title{Internal function to fit an empirical distribution}
\usage{
.getEmpiricalParam(d, title, verbose = FALSE, file.name = NULL)
}
\arguments{
\item{d}{A vector of values to fit.}

\item{title}{A plot title.}

\item{verbose}{Provide details on computations.}

\item{file.name}{The file name of a PDF file.}
}
\value{
A list with the step function implementing the CDF of
  the empirical distribution (\code{empirCDF}).

  If \code{file.name} is provided, a control plot is generated in a PDF with
  a data histogram and the fitted Gaussian. \code{title} is used to give this
  plot a main title.
}
\description{
Based on stats::ecdf.
}
\keyword{internal}
