% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.getGammaParam}
\alias{.getGammaParam}
\title{Internal function to fit a Gamma distribution}
\usage{
.getGammaParam(d, title, file.name = NULL)
}
\arguments{
\item{d}{A vector of values to fit.}

\item{title}{A plot title.}

\item{file.name}{The file name of a PDF file.}
}
\value{
A list with the shape (\code{k}) and scale (\code{theta}) estimates.

If \code{file.name} is provided, a control plot is generated in a PDF with a data histogram and
the fitted Gamma (plus a Gaussian for comparison). \code{title} is used to give this plot a main title.
}
\description{
Method of moments estimators are used.
}
