% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.getKernelEmpiricalParam}
\alias{.getKernelEmpiricalParam}
\title{Internal function to fit a Gaussian kernel-based empirical distribution}
\usage{
.getKernelEmpiricalParam(d, title, verbose = FALSE, file.name = NULL, n = 512)
}
\arguments{
\item{d}{A vector of values to fit.}

\item{title}{A plot title.}

\item{verbose}{Provide details on computations.}

\item{file.name}{The file name of a PDF file.}

\item{n}{The number of grid points for density FFT}
}
\value{
A list with the step function implementing the CDF of
  the empirical distribution (\code{kernelCDF}).

  If \code{file.name} is provided, a control plot is generated in a PDF with
  a data histogram and the fitted Gaussian. \code{title} is used to give this
  plot a main title.
}
\description{
Based on stats::density.
}
\keyword{internal}
