% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.getMixedGaussianParam}
\alias{.getMixedGaussianParam}
\title{Internal function to fit a censored mixed-Gaussian distribution}
\usage{
.getMixedGaussianParam(d, title, verbose = FALSE, file.name = NULL)
}
\arguments{
\item{d}{A vector of values to fit.}

\item{title}{A plot title.}

\item{verbose}{Provide details on computations.}

\item{file.name}{The file name of a PDF file.}
}
\value{
A list with the mean (\code{mu}) and standard deviation
  (\code{sigma}) estimates of each distribution along with the
  weight alpha applied to the first distribution.

  If \code{file.name} is provided, a control plot is generated in a PDF with
  a data histogram and the fitted Gaussian. \code{title} is used to give this
  plot a main title.
}
\description{
Maximum-likelihood estimators are used.
}
\keyword{internal}
