% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.pwHyperScores}
\alias{.pwHyperScores}
\title{Internal function to score pathways}
\usage{
.pwHyperScores(
  genes,
  sp2pw,
  fdr.proc = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BY",
    "ABH", "TSBH"),
  col.id,
  col.term,
  col.descr,
  max.pw.size = 200,
  min.pw.size = 5,
  min.inter = 2
)
}
\arguments{
\item{genes}{The set of differentialls expressed genes.}

\item{sp2pw}{A pathway reference table.}

\item{fdr.proc}{The procedure for adjusting P-values accoring to \code{\link[multtest]{mt.rawp2adjp}}.}

\item{col.id}{Column index or name in \code{pw} for the gene symbols.}

\item{col.term}{Column index or name in \code{pw} for the pathway IDs.}

\item{col.descr}{Column index or name in \code{pw} for the pathway names.}

\item{max.pw.size}{Maximum pathway size to consider from the pathway reference.}

\item{min.pw.size}{Minimum pathway size to consider from the pathway reference.}

\item{min.inter}{Minimum number of regulated genes in a pathway for this pathway to be reported.}
}
\value{
A table of pathways from the reference table \code{sp2pw} with different statistics and P-values.
}
\description{
Perform hypergeometric tests and collect intersection sizes.
}
