% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.readRDSFromCache}
\alias{.readRDSFromCache}
\title{Read RDS from the cache.}
\usage{
.readRDSFromCache(bfc, resourceName)
}
\arguments{
\item{bfc}{Object of class BiocFileCache, created by a call to
BiocFileCache::BiocFileCache()}

\item{resourceName}{keyword associated to a specific resourceName}
}
\value{
Returns BiocFileCache::bfcquery object or FALSE
}
\description{
Access  resources (pathways or network
stored in the cache.
}
\keyword{internal}
