% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.shufflePermutationIndices}
\alias{.shufflePermutationIndices}
\title{Internal function to shuffle permutation indices}
\usage{
.shufflePermutationIndices(pind, seed)
}
\arguments{
\item{pind}{Permutation indices such as returned by
\code{\link{.buildPermutationIndices}}.}

\item{seed}{Seed to reproduce exact same sampling.}
}
\value{
A list with same structure as \code{pind} with shuffled indices
  within each bin.
}
\description{
Internal function to shuffle permutation indices
}
