% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrepare.R
\name{.testCacheFiles}
\alias{.testCacheFiles}
\title{Check there is a well formated cache}
\usage{
.testCacheFiles()
}
\value{
Returns `NULL`, invisibly.
}
\description{
Check there is a well formated cache
}
