% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dualHeatmap}
\alias{dualHeatmap}
\title{Heatmap function for LR scores and additional data}
\usage{
dualHeatmap(
  mat.c,
  mat.e,
  width,
  height,
  file.name = NULL,
  dend.row = NULL,
  dend.spl = NULL,
  dend.e = NULL,
  cols = NULL,
  cols.e = NULL,
  pointsize = 4,
  cut.p = 0.01,
  vert.p = 0.9,
  row.names.1 = TRUE,
  row.names.2 = TRUE
)
}
\arguments{
\item{mat.c}{A matrix with the signature scores such as output by
\code{\link{scoreLRGeneSignatures}}.}

\item{mat.e}{A second matrix to feature below mat.c.}

\item{width}{PDF width.}

\item{height}{PDF height.}

\item{file.name}{A PDF file name.}

\item{dend.row}{A precomputed row dendrogram for \code{mat.c}.}

\item{dend.spl}{A precompute sample (column) dendrogram.}

\item{dend.e}{A precomputed row dendrogram for \code{mat.e}.}

\item{cols}{A vector of colors to use for the heatmap of
\code{mat.c}.}

\item{cols.e}{A vector of colors to use for the heatmap of
\code{mat.e}.}

\item{pointsize}{PDF pointsize.}

\item{cut.p}{Proportion of top and bottom values for thresholding.}

\item{vert.p}{Proportion of total height dedicated to the top heatmap.}

\item{row.names.1}{A logical to turn on/off the display of row (gene)
names in the top heatmap.}

\item{row.names.2}{A logical to turn on/off the display of row (gene)
names in the bottom heatmap.}
}
\value{
A heatmap. Since heatmap plotting tend to be slow on the screen, it
is advisable to provide a
PDF file name and plot in a file (much faster).

Extreme values (top and bottom) can be replaced by global quantiles at
\code{cut.p} and \code{1-cut.p}
to avoid color scales shrunk by a few outliers.

This is a convenience function that relies on the \code{ComplexHeatmap}
package to propose a simple way
of representing signature scores. If more advanced features are needed or
more graphic parameters
should be controlled, users should implement their own function.
}
\description{
Generate a stack of heatmaps. The top heatmap represents ligand-receptor
gene signature scores,
while the bottom heatmap represents the second, user-chosen data set to
feature along with gene signatures.
}
\examples{
\dontrun{
}
}
