% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getComplexes}
\alias{getComplexes}
\title{Retrieve LR complexes}
\usage{
getComplexes(idRelease = NULL)
}
\arguments{
\item{idRelease}{integer ID of the chosen release
Default is NULL such that the latest version is selected.}
}
\value{
Returns data frame Complexes, invisibly.
}
\description{
Fetch LR complexes from reference database and
and return a data frame
}
\examples{
getComplexes(idRelease=1)
}
