% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{getCorrelatedLR}
\alias{getCorrelatedLR}
\title{Get correlated ligand-receptor pairs}
\usage{
getCorrelatedLR(ds, min.cor = 0.3, restrict.genes = NULL)
}
\arguments{
\item{ds}{A BulkSignalR data set.}

\item{min.cor}{The minimum correlation required.}

\item{restrict.genes}{A list of gene symbols that restricts ligands and receptors.}
}
\value{
A list containing a table of putative ligand-receptor pairs along with
their correlations above \code{min.cor}. This list is the first step of a ligand-receptor analysis.

The \code{restrict.genes}
parameter is used for special cases where LRdb must be further restricted to a subset.
}
\description{
Compute the Spearman correlations of all the ligand-receptor pairs in LRdb and return those above a minimum value.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds.LR <- getCorrelatedLR(ds)
}
}
