% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getInteractions}
\alias{getInteractions}
\title{Retrieve LR interactions.}
\usage{
getInteractions(idRelease = NULL)
}
\arguments{
\item{idRelease}{integer ID of the chosen release
Default is NULL such that the latest version is selected.}
}
\value{
Returns data frame LR interactions, invisibly.
}
\description{
Fetch LR interactions from reference database and
and return a data frame
}
\examples{
getInteractions(idRelease=1)
}
