% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{getLRGeneSignatures,BSRInference-method}
\alias{getLRGeneSignatures,BSRInference-method}
\title{Extract gene signatures of LR pair activity}
\usage{
\S4method{getLRGeneSignatures}{BSRInference}(obj, pval.thres = NULL, qval.thres = NULL)
}
\arguments{
\item{pval.thres}{P-value threshold.}

\item{qval.thres}{Q-value threshold.}
}
\value{
A BSRSignature object containing a gene signature for each triple
ligand-receptor pair. A reduction to the best pathway
for each pair is automatically performed and the gene signature is
comprised of the ligand, the receptor,
and all the target genes with rank equal or superior to \code{pairs$rank}.
In case \code{signed==TRUE},
the rank is defined for correlation absolute values.
}
\description{
Obtains gene signatures reflecting ligand-receptor as well as
receptor downstream activity to
score ligand-receptor pairs across samples subsequently with
\code{\link{scoreLRGeneSignatures}}.
}
\examples{
\dontrun{
}
}
