% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{getLRGeneSignatures}
\alias{getLRGeneSignatures}
\title{Extract gene signatures of LR pair activity}
\usage{
getLRGeneSignatures(
  pairs,
  LLR.thres = NULL,
  pval.thres = NULL,
  qval.thres = NULL,
  signed = TRUE
)
}
\arguments{
\item{pairs}{A ligand-receptor table such as output by \code{pValuesLR} and \code{naiveBayesLR}.}

\item{LLR.thres}{Log-likelihood threshold.}

\item{pval.thres}{P-value threshold.}

\item{qval.thres}{Q-value threshold.}

\item{signed}{A logical indicating whether correlations were be considered with their signs upon P-value or LLR estimations.}
}
\value{
A table with a gene signature for each triple ligand-receptor pair. A reduction to the best pathway
for each pair is automatically performed and the gene signature is comprised of the ligand, the receptor,
and all the target genes with rank equal or superior to \code{pairs$rank}. In case \code{signed==TRUE},
the rank is defined for correlation absolute values.
}
\description{
Obtains gene signatures reflecting ligand-receptor as well as receptor downstream activity to
score ligand-receptor pairs across samples subsequently with \code{scoreLRGeneSignatures()}.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pairs.p <- pValuesLR(ds.LR,ds$param)
signatures <- getPathwayStats(pairs.p,qval.thres=0.01)
}
}
