% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{getLRNetwork}
\alias{getLRNetwork}
\title{Generate a ligand-receptor network}
\usage{
getLRNetwork(
  pairs,
  LLR.thres = NULL,
  pval.thres = NULL,
  qval.thres = NULL,
  node.size = 5
)
}
\arguments{
\item{pairs}{A ligand-receptor table such as output by \code{pValuesLR} and \code{naiveBayesLR}.}

\item{LLR.thres}{Log-likelihood threshold.}

\item{pval.thres}{P-value threshold.}

\item{qval.thres}{Q-value threshold.}

\item{node.size}{Default node size in the network.}
}
\value{
An \code{igraph} object featuring the ligand-receptor network. Default colors and node sizes are assigned,
which can be changed afterwards if necessary.
}
\description{
Generate a ligand-receptor network from a ligand-receptor table.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pp <- pValuesLR(ds.LR,ds$param)
gLR <- getLRNetwork(pp,qval.thres=0.01)
plot(gLR)
write.graph(gLR,file="SDC-LR-network.graphml",format="graphml")
}
}
