% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{getMultipleLRIntracellNetworks}
\alias{getMultipleLRIntracellNetworks}
\title{Generate multiple ligand-receptor-downstrem signaling networks}
\usage{
getMultipleLRIntracellNetworks(
  ds,
  pairs,
  n.clusters,
  min.score = 0,
  LLR.thres = NULL,
  pval.thres = NULL,
  qval.thres = NULL,
  cut.p = 0.01,
  min.cor = 0.3,
  signed = FALSE,
  restrict.pw = NULL,
  node.size = 5
)
}
\arguments{
\item{ds}{A BulkSignalR data set.}

\item{pairs}{A ligand-receptor table such as output by \code{pValuesLR} and \code{naiveBayesLR}.}

\item{n.clusters}{The number of clusters.}

\item{min.score}{The minimum required average z-score of a gene signature in a cluster.}

\item{LLR.thres}{Log-likelihood threshold in \code{pairs}.}

\item{pval.thres}{P-value threshold in \code{pairs}.}

\item{qval.thres}{Q-value threshold in \code{pairs}.}

\item{cut.p}{Proportion of top and bottom values for thresholding.}

\item{min.cor}{Minimum correlation required for the target genes.}

\item{signed}{A logical indicating whether \code{min.cor} is imposed to correlation absolute values (\code{FALSE}) or not (\code{TRUE}).}

\item{restrict.pw}{A vector of pathway IDs to which receptor downstream signaling is restricted.}

\item{node.size}{Default node size in the network.}
}
\value{
A list containing the main elements of the clustering analysis as well as an inner list
containing the \code{igraph} objects created for each cluster.

The elements of clustering are the retained global ligand-receptor table, the correspondinf gene signatures and their
scores across \code{ds} samples, the resulting hierarchical cultering and sample membership.
}
\description{
For each cluster of samples based on ligand-receptor gene signatures, generate a ligand-receptor network
from a ligand-receptor table and add the shortest paths from the receptor to correlated
target genes following Reactome and KEGG pathways.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pp <- pValuesLR(ds.LR,ds$param)
mult.intra.net <- getMultipleLRNetworks(ds,pp,n.clusters=4,qval.thres=0.01)
plot(mult.intra.net$hclust.spl)
lay <- layout_with_kk(mult.intra.net$networks[[1]])
plot(mult.intra.net$networks[[1]],layout=lay)
}
}
