% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{getMultipleLRNetworks}
\alias{getMultipleLRNetworks}
\title{Generate multiple ligand-receptor networks}
\usage{
getMultipleLRNetworks(
  ds,
  pairs,
  n.clusters,
  min.score = 0,
  pval.thres = NULL,
  qval.thres = NULL,
  cut.p = 0.01,
  node.size = 5
)
}
\arguments{
\item{ds}{A BulkSignalR data set.}

\item{pairs}{A ligand-receptor table such as output by \code{pValuesLR}.}

\item{n.clusters}{The number of clusters.}

\item{min.score}{The minimum required average z-score of a gene signature in a cluster.}

\item{pval.thres}{P-value threshold in \code{pairs}.}

\item{qval.thres}{Q-value threshold in \code{pairs}.}

\item{cut.p}{Proportion of top and bottom values for thresholding.}

\item{node.size}{Default node size in the network.}
}
\value{
A list containing the main elements of the clustering analysis as well as an inner list
containing the \code{igraph} objects created for each cluster.

The elements of clustering are the retained global ligand-receptor table, the corresponding gene signatures and their
scores across \code{ds} samples, the resulting hierarchical clustering and sample membership.
}
\description{
Generate a ligand-receptor network for each cluster of samples based on ligand-receptor gene signatures.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pp <- pValuesLR(ds.LR,ds$param)
gLR <- getLRNetwork(pp,qval.thres=0.01)
mult.net <- getMultipleLRNetworks(ds,pp,n.clusters=4,qval.thres=0.01)
plot(mult.net$hclust.spl)
lay <- layout_with_kk(mult.net$networks[[1]])
plot(mult.net$networks[[1]],layout=lay)
}
}
