% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{getPathwayStats}
\alias{getPathwayStats}
\title{Basic statistics about hit pathways}
\usage{
getPathwayStats(pairs, LLR.thres = NULL, pval.thres = NULL, qval.thres = NULL)
}
\arguments{
\item{pairs}{A ligand-receptor table such as output by \code{pValuesLR} and \code{naiveBayesLR}.}

\item{LLR.thres}{Log-likelihood threshold.}

\item{pval.thres}{P-value threshold.}

\item{qval.thres}{Q-value threshold.}
}
\value{
A table with the pathways selected after the chosen threshold was applied to rows in \code{pairs}.
Each pathway is reported along with various statistics:
the number of selected receptors in this pathway, the total number of receptors described this pathway,
the number of selected ligand-receptor pairs hitting this pathway, and the total number of
ligand-receptor pairs described taht could hit this pathway.

Obviously, one could imagine computing enrichments in receptors or ligand-receptor pairs
based on such statistics, but the actual meaning of such an analysis would be ambiguous since
the pathways were already selected as significantly regulated by the receptor. We thus did not implement
this (elementary) computation.
}
\description{
Basic statistics about hit pathways
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pairs.p <- pValuesLR(ds.LR,ds$param)
pw.stat <- getPathwayStats(pairs.p,qval.thres=0.01)
}
}
