% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{getResource}
\alias{getResource}
\title{Get resources from the cache}
\usage{
getResource(resourceName = NULL, cache = FALSE)
}
\arguments{
\item{resourceName}{Resource name.}

\item{cache}{Logical. Default is FALSE
If TRUE, you will use environment variables.}
}
\value{
Returns a data frame containing the requested
resource.
}
\description{
Get resources (pathways, or PathwayCommons network
from \url{https://www.pathwaycommons.org/})
stored in the cache.
}
\examples{
reactome <- getResource(resourceName = "Reactome",cache=TRUE)
}
