% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immune.signatures}
\alias{immune.signatures}
\title{Immune cell gene signatures}
\format{
A data frame with 1541 rows and 2 variables:
\describe{
  \item{gene}{HUGO gene symbol}
  \item{signature}{cell population name}
}
}
\source{
PanglaoDB (Franzén et al., Database, 2019).
}
\usage{
data(immune.signatures)
}
\description{
A dataset containing gene signatures for general immune cell populations.
}
\keyword{datasets}
