% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{inferenceParameters<-,BSRInference-method}
\alias{inferenceParameters<-,BSRInference-method}
\title{Inference parameters setter (internal use only)}
\usage{
\S4method{inferenceParameters}{BSRInference}(x) <- value
}
\arguments{
\item{x}{BRSInference object.}

\item{value}{value to be set.}
}
\value{
returns \code{NULL}
}
\description{
Inference parameters setter (internal use only)
}
\keyword{internal}
