% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{inferenceParameters}
\alias{inferenceParameters}
\alias{inferenceParameters,BSRInference-method}
\title{Inference parameters accessor}
\usage{
\S4method{inferenceParameters}{BSRInference}(x)
}
\arguments{
\item{x}{BRSInference object.}
}
\value{
inf.param
}
\description{
Inference parameters accessor
}
\examples{
bsrinf <- new ("BSRInference")
inferenceParameters(bsrinf)
}
