% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{infParam,BSRInference-method}
\alias{infParam,BSRInference-method}
\title{Inference parameters accessor}
\usage{
\S4method{infParam}{BSRInference}(x)
}
\description{
Inference parameters accessor
}
