% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{infParam<-,BSRInference-method}
\alias{infParam<-,BSRInference-method}
\title{Inference parameters setter (internal use only)}
\usage{
\S4method{infParam}{BSRInference}(x) <- value
}
\arguments{
\item{x}{BRSInferecence object.}

\item{value}{value to be set.}
}
\description{
Inference parameters setter (internal use only)
}
\keyword{internal}
