% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{infParam}
\alias{infParam}
\alias{infParam,BSRInference-method}
\title{Inference parameters accessor}
\usage{
\S4method{infParam}{BSRInference}(x)
}
\arguments{
\item{x}{BRSInference object.}
}
\value{
infParam
}
\description{
Inference parameters accessor
}
\examples{
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))

bsrdm <- prepareDataset(sdc[,-normal])
bsrdm <- learnParameters(bsrdm, quick=FALSE,
plot.folder=".", verbose=TRUE)
bsrinf <- initialInference(bsrdm)
infParam(bsrinf)
}
