% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{initialOrganism,BSRDataModel-method}
\alias{initialOrganism,BSRDataModel-method}
\title{organism accessor}
\usage{
\S4method{initialOrganism}{BSRDataModel}(x)
}
\description{
organism accessor
}
