% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{initialOrganism}
\alias{initialOrganism}
\alias{initialOrganism,BSRDataModel-method}
\title{organism accessor}
\usage{
\S4method{initialOrganism}{BSRDataModel}(x)
}
\arguments{
\item{x}{Object BSRDataModel}
}
\value{
initialOrganism
}
\description{
organism accessor
}
\examples{
bsrdm <- BSRDataModel(
    counts = matrix(1.5,
                nrow = 2, ncol = 2,
                dimnames = list(c("A", "B"), c("C", "D"))
                   ),
    method = "TC",
    log.transformed = TRUE,
    normalize = FALSE,
    min.LR.found = 0
)
initialOrganism(bsrdm)
}
