% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{initialOrthologs,BSRDataModel-method}
\alias{initialOrthologs,BSRDataModel-method}
\title{Model parameter accessor}
\usage{
\S4method{initialOrthologs}{BSRDataModel}(x)
}
\description{
Model parameter accessor
}
