% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{learnParameters}
\alias{learnParameters}
\title{Training of BulkSignalR model parameters}
\usage{
learnParameters(
  ds,
  normal = NULL,
  induced = NULL,
  plot.folder = NULL,
  verbose = FALSE,
  n.rand.LR = 5,
  n.rand.RT = 2,
  with.complex = TRUE,
  max.pw.size = 200,
  min.pw.size = 5,
  min.positive = 4,
  quick = TRUE,
  force.altern = FALSE
)
}
\arguments{
\item{ds}{A BulkSignalR data set, i.e., a list with the normalized read counts and sample types.}

\item{normal}{A character vector defining the sample types that correspond to the "normal" condition in case alternative distributions are learnt.}

\item{induced}{A character vector defining the sample types that correspond to the "induced" condition in case alternative distributions are learnt.}

\item{plot.folder}{A folder name for generating control plots.}

\item{verbose}{A logical activating progress messages for the user.}

\item{n.rand.LR}{The number of random expression matrices to use for learning the ligand-receptor correlation distribution.}

\item{n.rand.RT}{The number of random expression matrices to use for learning the receptor-target genes correlation distribution.}

\item{with.complex}{A logical indicating whether receptor co-complex members should be included in the target genes.}

\item{max.pw.size}{Maximum pathway size to consider from the pathway reference.}

\item{min.pw.size}{Minimum pathway size to consider from the pathway reference.}

\item{min.positive}{Minimum number of target genes to be found in a given pathway.}

\item{quick}{A logical indicating whether approximate parameters for the receptor-target correlations should be used.}

\item{force.altern}{A logical to force computations when learning the alternative distributions on a limited number of data.}
}
\value{
The BulkSignalR data set \code{ds} extended with a slot \code{ds$param} that contains all the null model parameters
as well as parameters that were used to specify how they were learnt. Provided \code{induced} and \code{normal} are set, the
alternative model parameters are also added into \code{ds$param}. \code{ds$param} is a multi-level named list.

In a reference pathway, i.e., a Reactome pathway or the genes of a GOBP term, the target genes are the
genes coding for proteins forming a complex with the receptor and the genes in the pathway downstream the receptor,
which are given as regulated by the pathway. If \code{with.complex} is set to \code{FALSE}, then only the
regulated genes are considered. Participation to a complex and being regulated as well as the pathway directed topologies
are defined by Reactome and KEGG pathways as provided by PathwayCommons. Those parameters should be identical to the
values intended when searching for ligand-receptor pairs with \code{\link{getCorrelatedLR}} and
\code{\link{checkReceptorSignaling}}. Although the statistical distributions are rather robust it is not advisable
to use different parameters that could nonetheless introduce biases, but for saving compute time and exploring.

The maximum pathway size is used to limit the redundancy inherent to GOBP and Reactome. The minimum pathway size is
used to avoid overspecific, noninformative results.

If alternative distributions are learnt, then the \code{induced} and \code{normal} conditions
define the samples involved in the comparison "normal" versus "induced". To train alternative ditribution parameters is necessary
in case one wants to use log-likelihood ratios (LLR) to score ligand-receptor pairs (\code{\link{naiveBayesLR}}) instead of simpler P-value estimations
(\code{\link{pValuesLR}}) based on the sole null distributions. Such a training relies on differential pathways
found in the "normal" versus "induced" differential gene analysis (uning edgeR).

Note that to be coherent, in case alternative distributions are learnt, then the null model is learnt on the samples
in the induced condition only (all the samples otherwise). This logic is motivated by a classical
scenario where one wants to find ligand-receptor pairs in diseased samples and one has access to expression data for
both normal and diseased samples. The comparison normal versus disease is obviously used to find deregulated pathways,
and the condition of interest to score the ligand-receptor pairs is diseased.

If the alternative distributions are not learnt, then \code{normal} and \code{induced} should not be used.

The parameter \code{force.altern} should not be used in general.
It forces the estimation of the alternative model parameters even in cases where very limited numbers of differential pathways are available. There
is a risk of error executing \code{learnParameters} with this parameter set to \code{TRUE}. Use for data exploration only.
}
\description{
Unique entry point for training the parameters behind BulkSignalR statistical models.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")
}
}
