% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{ligands<-,BSRInference-method}
\alias{ligands<-,BSRInference-method}
\title{ligands setter (internal use only)}
\usage{
\S4method{ligands}{BSRInference}(x) <- value
}
\description{
ligands setter (internal use only)
}
