% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R, R/BSRSignature-class.R
\name{ligands}
\alias{ligands}
\alias{ligands,BSRInference-method}
\alias{ligands,BSRSignature-method}
\title{ligands accessor}
\usage{
\S4method{ligands}{BSRInference}(x)

\S4method{ligands}{BSRSignature}(x)
}
\arguments{
\item{x}{BSRSignature}
}
\value{
ligands

ligands
}
\description{
ligands accessor

ligands accessor
}
\examples{
bsr.sig <- new("BSRSignature")
ligands(bsr.sig)
}
