% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{logTransformed}
\alias{logTransformed}
\alias{logTransformed,BSRDataModel-method}
\title{log.transformed accessor}
\usage{
\S4method{logTransformed}{BSRDataModel}(x)
}
\arguments{
\item{x}{Object BRSDataModel}
}
\value{
logTransformed
}
\description{
log.transformed accessor
}
\examples{
bsrdm <- BSRDataModel(
    counts = matrix(1.5,
                nrow = 2, ncol = 2,
                dimnames = list(c("A", "B"), c("C", "D"))
                   ),
    method = "TC",
    log.transformed = TRUE,
    normalize = FALSE,
    min.LR.found = 0
)
logTransformed(bsrdm)
}
