% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{LRinter<-,BSRInference-method}
\alias{LRinter<-,BSRInference-method}
\title{LRinter setter (internal use only)}
\usage{
\S4method{LRinter}{BSRInference}(x) <- value
}
\description{
LRinter setter (internal use only)
}
