% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{LRinter}
\alias{LRinter}
\alias{LRinter,BSRInference-method}
\title{LRinter accessor}
\usage{
\S4method{LRinter}{BSRInference}(x)
}
\arguments{
\item{x}{BSRInference object}
}
\value{
LRinter
}
\description{
LRinter accessor
}
\examples{
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))

bsrdm <- prepareDataset(sdc[,-normal])
bsrdm <- learnParameters(bsrdm, quick=FALSE,
plot.folder=".", verbose=TRUE)
bsrinf <- initialInference(bsrdm)
LRinter(bsrinf)
}
