% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R,
%   R/BSRInferenceComp-class.R
\name{LRinterShort}
\alias{LRinterShort}
\alias{LRinterShort,BSRInference-method}
\alias{LRinterShort,BSRInferenceComp-method}
\title{Simplified LRinter accessor reporting the essential columns}
\usage{
\S4method{LRinterShort}{BSRInference}(x)

\S4method{LRinterShort}{BSRInferenceComp}(x)
}
\arguments{
\item{x}{BSRInferenceComp object}
}
\value{
LRinterShort

LRinterShort
}
\description{
Simplified LRinter accessor reporting the essential columns

Simplified LRinter accessor reporting the essential columns
}
\examples{
data(bsrinf.comp, package = "BulkSignalR")
LRinterShort(bsrinf.comp)[5,]
}
