% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModelComp-class.R
\name{mu<-,BSRDataModelComp-method}
\alias{mu<-,BSRDataModelComp-method}
\title{Mu setter (internal use only)}
\usage{
\S4method{mu}{BSRDataModelComp}(x) <- value
}
\arguments{
\item{x}{object BSRDataModelComp}

\item{value}{value to be set for BSRDataModelComp}
}
\value{
returns \code{NULL}
}
\description{
Mu setter (internal use only)
}
\keyword{internal}
