% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{ncounts,BSRDataModel-method}
\alias{ncounts,BSRDataModel-method}
\title{Normalized count matrix accessor}
\usage{
\S4method{ncounts}{BSRDataModel}(x)
}
\description{
Normalized count matrix accessor
}
