% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{ncounts}
\alias{ncounts}
\alias{ncounts,BSRDataModel-method}
\title{Normalized count matrix accessor}
\usage{
\S4method{ncounts}{BSRDataModel}(x)
}
\arguments{
\item{x}{object BSRDataModel}
}
\value{
ncounts
}
\description{
Normalized count matrix accessor
}
\examples{
bsrdm <- BSRDataModel(
    counts = matrix(1.5,
                nrow = 2, ncol = 2,
                dimnames = list(c("A", "B"), c("C", "D"))
                   ),
    method = "TC",
    log.transformed = TRUE,
    normalize = FALSE,
    min.LR.found = 0
)
ncounts(bsrdm)
}
