% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{normalization,BSRDataModel-method}
\alias{normalization,BSRDataModel-method}
\title{Normalization accessor}
\usage{
\S4method{normalization}{BSRDataModel}(x)
}
\description{
Normalization accessor
}
