% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ortholog.dict}
\alias{ortholog.dict}
\title{A skinny data frame used in the mouse workflow}
\format{
An example of a data frame created by findOrthoGenes
}
\usage{
data(ortholog.dict)
}
\description{
Synthetic object used during the call to the
function `resetToInitialOrganism``
}
\keyword{datasets}
