% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p.EMT}
\alias{p.EMT}
\title{Partial EMT gene signature}
\format{
A data frame with 100 rows and 1 variables:
\describe{
  \item{gene}{HUGO gene symbol}
}
}
\source{
Puram, SV & al., Cell, 2017.
}
\usage{
data(p.EMT)
}
\description{
A dataset containing a partial EMT gene signature.
}
\keyword{datasets}
