% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{parameters<-,BSRDataModel-method}
\alias{parameters<-,BSRDataModel-method}
\title{Parameters dataModel setter (internal use only)}
\usage{
\S4method{parameters}{BSRDataModel}(x) <- value
}
\arguments{
\item{x}{object BSRDataModel}

\item{value}{value to be set for BSRDataModel}
}
\value{
returns \code{NULL}
}
\description{
Parameters dataModel setter (internal use only)
}
\keyword{internal}
