% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{pValuesLR}
\alias{pValuesLR}
\title{Ligand-receptor pair P-value estimation}
\usage{
pValuesLR(
  lr,
  param,
  rank.p = 0.75,
  signed = TRUE,
  fdr.proc = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BY",
    "ABH", "TSBH"),
  use.gamma = FALSE,
  best.pw.only = FALSE
)
}
\arguments{
\item{lr}{A ligand-receptor analysis such as output by \code{checkReceptorSignaling}.}

\item{param}{A list containing the statistical model parameters.}

\item{rank.p}{A number between 0 and 1 defining the rank of the last considered target genes.}

\item{signed}{A logical indicating whether correlations should be considered with their signs.}

\item{fdr.proc}{The procedure for adjusting P-values according to \code{\link[multtest]{mt.rawp2adjp}}.}

\item{use.gamma}{A logical indicating whether the Gamma distribution should be used instead of the Gaussian.}

\item{best.pw.only}{A logical indicating that only the pathway with best P-value should be reported for each ligand-receptor pair.}
}
\value{
A table reporting every ligand-receptor pair with P-values and Q-values (adjusted P-values).
}
\description{
Estimate the P-value of each ligand-receptor pair in the \code{merged.pairs} slot of the
ligand-receptor analysis \code{lr}.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pairs.p <- pValuesLR(ds.LR,ds$param)
}
}
