% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{receptors<-,BSRInference-method}
\alias{receptors<-,BSRInference-method}
\title{receptors setter (internal use only)}
\usage{
\S4method{receptors}{BSRInference}(x) <- value
}
\arguments{
\item{x}{BRSInference object}

\item{value}{value to be set for BRSInference}
}
\value{
returns \code{NULL}
}
\description{
receptors setter (internal use only)
}
\keyword{internal}
