% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R, R/BSRSignature-class.R
\name{receptors}
\alias{receptors}
\alias{receptors,BSRInference-method}
\alias{receptors,BSRSignature-method}
\title{receptors accessor}
\usage{
\S4method{receptors}{BSRInference}(x)

\S4method{receptors}{BSRSignature}(x)
}
\arguments{
\item{x}{BSRSignature}
}
\value{
receptors

receptors
}
\description{
receptors accessor

receptors accessor
}
\examples{
bsr.sig <- new("BSRSignature")
ligands(bsr.sig)
}
