% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{reduceToBestPathway,BSRInference-method}
\alias{reduceToBestPathway,BSRInference-method}
\title{Keep one pathway per ligand-receptor pair}
\usage{
\S4method{reduceToBestPathway}{BSRInference}(obj)
}
\value{
A BSRInference object reduced to only report one pathway per
ligand-receptor pair. The pathway with the
smallest P-value is selected.
}
\description{
Keep one pathway per ligand-receptor pair
}
\details{
During the execution of \code{pValuesLR}, ligand-receptor pairs
are evaluated in relation with pathways that allow checking receptor
downstream correlations. It is thus possible
that several pathways are reported for a same LR pair.
}
\examples{
print('reduceToBestPathway')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
bsrinf.redBP  <- reduceToBestPathway(bsrinf)  

}
