% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{reduceToBestPathway}
\alias{reduceToBestPathway}
\title{Keep one pathway per ligand-receptor pair}
\usage{
reduceToBestPathway(pairs)
}
\arguments{
\item{pairs}{A ligand-receptor table such as output by \code{pValuesLR} and \code{naiveBayesLR}.}
}
\value{
A table with the same structure as \code{pairs} but reduced to only report one pathway by ligand-receptor pair.
Depending on how the pairs were scored, \code{pValuesLR} or \code{naiveBayesLR}, the pathway with
smallest P-value or largest log-likelihood ratio is selected.

During the execution of \code{pValuesLR} or \code{naiveBayesLR}, ligand-receptor pairs are evaluated
in relation with pathways that allow checking receptor downstream correlations. It is thus possible
that several pathways are reported for a same pair.
}
\description{
Keep one pathway per ligand-receptor pair
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pairs.p <- pValuesLR(ds.LR,ds$param)
pairs.red <- reduceToBestPathway(pairs.p)
}
}
