% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{reduceToLigand,BSRInference-method}
\alias{reduceToLigand,BSRInference-method}
\title{Aggregate the receptors of a same ligand}
\usage{
\S4method{reduceToLigand}{BSRInference}(obj)
}
\value{
A BSRInference object but reduced to one row per ligand.
All the receptors are combined in a
semi-colon-separated list surrounded by curly braces in the tabular
slot \code{LRinter}, and in vectors in the \code{ligands} (list) slot.

The reported P-value and target genes are those from the pathway with
the smallest P-value.
}
\description{
Simplifies a ligand-receptor table to focus on the ligands.
}
\examples{
\dontrun{
bsrinf.redL <- reduceToLigand(bsrinf)
}
}
