% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{reduceToLigand}
\alias{reduceToLigand}
\title{Aggregate the receptors of a same ligand}
\usage{
reduceToLigand(pairs)
}
\arguments{
\item{pairs}{A ligand-receptor table such as output by \code{pValuesLR} and \code{naiveBayesLR}.}
}
\value{
A table with the same structure as \code{pairs} but reduced to only report one row per ligand.
All the receptors are combined in a semi-colon-separated list surounded by curly braces.

To limit the output table to a single line per ligand, this function also selects the pathway
with the smallest P-value, respectively the largest log-likelihood ratio. The reported list of target genes
and their respective correlations are those of this single selected pathway.
}
\description{
Simplifies a ligand-receptor table to focus on the ligands.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pairs.p <- pValuesLR(ds.LR,ds$param)
pairs.lig <- reduceToLigand(pairs.p)
}
}
