% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{reduceToPathway,BSRInference-method}
\alias{reduceToPathway,BSRInference-method}
\title{Aggregate ligands and receptors at the pathway level

Simplifies a ligand-receptor inference object to focus on the pathways.}
\usage{
\S4method{reduceToPathway}{BSRInference}(obj)
}
\value{
A BSRInference object reduced to only report one row per pathway.
The information of which ligand interacted with which receptor is lost as
all the ligands and all the receptors forming pairs related to a certain
pathway are combined.
For a given pathway, the reported P-values and target genes are those of
the best ligand-receptor pair that
was in this pathway.
Receptors and ligands are combined in two semi-colon-separated
lists surrounded by curly braces in the tabular slot \code{LRinter},
while the list representation slots (\code{ligands} and
\code{receptors}) are update accordingly.
}
\description{
Aggregate ligands and receptors at the pathway level

Simplifies a ligand-receptor inference object to focus on the pathways.
}
\examples{
\dontrun{
bsrinf.redP <- reduceToPathway(bsrinf)
}
}
