% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{reduceToReceptor,BSRInference-method}
\alias{reduceToReceptor,BSRInference-method}
\title{Aggregate the ligands of a same receptor

Simplifies a ligand-receptor table to focus on the receptors.}
\usage{
\S4method{reduceToReceptor}{BSRInference}(obj)
}
\value{
BSRInference object reduced to one row per receptor.
All the ligands are combined in a
semi-colon-separated list surrounded by curly braces in the tabular
slot \code{LRinter}, and in vectors in the \code{ligands} (list) slot.

The reported P-value and target genes are those from the line with the
pathway featuring the smallest P-value.
}
\description{
Aggregate the ligands of a same receptor

Simplifies a ligand-receptor table to focus on the receptors.
}
\examples{
\dontrun{
bsrinf.redR <- reduceToReceptor(bsrinf)
}
}
