% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resetPathwaysFromFile}
\alias{resetPathwaysFromFile}
\title{Import pathways from a file to dataframe}
\usage{
resetPathwaysFromFile(
  file,
  fileType = c("json", "gmt", "txt"),
  resourceName = NULL
)
}
\arguments{
\item{file}{Path to file.}

\item{fileType}{Default is Json.
Other options are gmt or txt files.}

\item{resourceName}{Two options "GO-BP" or "REACTOME".}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Pathways are defined in Reactome and
GoBP databases.
Those can be updated using
json files from
the Human Molecular Signatures Database (MSigDB)
at \url{https://www.gsea-msigdb.org/}
}
\details{
\code{resetDownstreamPathways} is a function
we provide to user to refresh REACTOME
and GO-BP content included in BulkSignalR.

Gmt file format also can be imported.
}
\examples{
print("resetPathwaysFromFile")
if (FALSE) {
    resetPathwaysFromFile(file, "GO-BP")
}

}
