% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scalesReverse}
\alias{scalesReverse}
\title{Flip over horizontal/vertical axis}
\usage{
scalesReverse(rasterImage, scale = "x")
}
\arguments{
\item{rasterImage}{raster image}

\item{scale}{Reverse horizontal(x)/vertical(y) axis}
}
\value{
A flipped raster image
}
\description{
Flip over horizontal/vertical axis
}
\examples{
print('scalesReverse')
if(FALSE){
img <- png::readPNG(path.to.file)
my.image.as.raster <- grDevices::as.raster(path.to.file)  
my.image.flipped <- scalesReverse(my.image.as.raster,scale="x")

 }
}
