% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{scoreLRGeneSignatures,BSRDataModel-method}
\alias{scoreLRGeneSignatures,BSRDataModel-method}
\title{Score ligand-receptor gene signatures

Compute ligand-receptor gene signature scores over a BSRDataModel.}
\usage{
\S4method{scoreLRGeneSignatures}{BSRDataModel}(
  obj,
  sig,
  LR.weight = 0.5,
  robust = FALSE,
  rename.by.pathway = FALSE,
  abs.z.score = FALSE
)
}
\arguments{
\item{sig}{A BSRSignature object.}

\item{LR.weight}{A number between 0 and 1 defining the relative weight
of the ligand and the receptor in the signature.}

\item{robust}{A logical indicating that z-scores should be computed
with median and MAD instead of mean and standard deviation.}

\item{rename.by.pathway}{A logical indicating whether row names of the
resulting score matrix should be pathway names instead of ligand-receptor
gene symbols.}

\item{abs.z.score}{A logical to use absolute z-scores (useful if the
activity of a paythway is reported by a mixture of up- and down-genes
whose z-score averages might hide actual activity).}
}
\value{
A matrix containing the scores of each ligand-receptor gene
signature in each sample.
}
\description{
Score ligand-receptor gene signatures

Compute ligand-receptor gene signature scores over a BSRDataModel.
}
\examples{
\dontrun{
}
}
