% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scoreLRGeneSignatures}
\alias{scoreLRGeneSignatures}
\title{Score ligand-receptor gene signatures}
\usage{
scoreLRGeneSignatures(
  ds,
  sig,
  sample.types = NULL,
  LR.weight = 0.5,
  robust = FALSE,
  rename.by.pathway = FALSE
)
}
\arguments{
\item{ds}{A BulkSignalR data set.}

\item{sig}{Gene signatures computed by \code{getLRGeneSignatures()}.}

\item{sample.types}{A character vector defining which samples in \code{ds} should be used.}

\item{LR.weight}{A number between 0 and 1 defining the relative weight of the ligand and the receptor in the signature.}

\item{robust}{A logical indicating that z-scores should be computed with median and MAD instead of mean and standard deviation.}

\item{rename.by.pathway}{A logical indicating whether row names of the result score matrix should be pathway names instead of ligand-receptor gene symbols.}
}
\value{
A matrix containing the scores of each ligand-receptor gene signature in each sample.

A selection of samples can be obtained by using the \code{sample.types} parameter.
}
\description{
Compute ligand-receptor gene signature scores over a BulkSIgnalR data set.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pairs.p <- pValuesLR(ds.LR,ds$param)
p.red.P <- reduceToPathway(pairs.p)
signatures <- getPathwayStats(p.red.P,qval.thres=0.01)
scores <- scoreLRGeneSignatures(ds,signatures,rename.by.pathway=TRUE)
}
}
