% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scoreSignatures}
\alias{scoreSignatures}
\title{Gene signature scoring}
\usage{
scoreSignatures(ds, ref.signatures, sample.types = NULL, robust = FALSE)
}
\arguments{
\item{ds}{A BulkSignalR data set.}

\item{ref.signatures}{Gene signatures.}

\item{sample.types}{A character vector defining which samples in \code{ds} should be used.}

\item{robust}{A logical indicating that z-scores should be computed with median and MAD instead of mean and standard deviation.}
}
\value{
A matrix containing the scores of each gene signature in each sample. Note that ligand-receptor gene
signature scores should be computed with \code{scoreLRGeneSignatures()} instead.

A selection of samples can be obtained by using the \code{sample.types} parameter.
}
\description{
Scores generic gene signatures over the samples of a BulkSignalR data set.
}
\examples{
\dontrun{
tme.signatures <- fread("TME-population-signatures.txt",data.table=FALSE)
tme.scores <- scoreSignatures(ds,tme.signatures)
}
}
