% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R
\name{sourceComparisonName}
\alias{sourceComparisonName}
\alias{sourceComparisonName,BSRInferenceComp-method}
\title{Source comparison name accessor}
\usage{
\S4method{sourceComparisonName}{BSRInferenceComp}(x)
}
\arguments{
\item{x}{BSRInferenceComp object}
}
\value{
src.comp.name
}
\description{
Source comparison name accessor
}
\examples{
bsrinf <- new("BSRInferenceComp")
sourceComparisonName(bsrinf)
}
