% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatialAssociation}
\alias{spatialAssociation}
\title{Statistical association of scores with area labels}
\usage{
spatialAssociation(
  scores,
  areas,
  test = c("Kruskal-Wallis", "ANOVA"),
  label.col = "label",
  idSpatial.col = "idSpatial",
  fdr.proc = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BY", "ABH",
    "TSBH")
)
}
\arguments{
\item{scores}{A matrix of scores, one L-R interaction per row and
spatial locations in the columns. This matrix is typically obtained
from BulkSignalR functions \code{scoreLRGeneSignatures} or \code{scScoring}.}

\item{areas}{A data.frame containing at least the x and y
coordinates of the locations, the unique IDs of spatial locations, and
a label column.}

\item{test}{The chosen statistical test, parametric (normal) or
nonparametric (see also details below).}

\item{label.col}{Column name in \code{areas} containing area labels.}

\item{idSpatial.col}{Column name in \code{areas} containing the unique
IDs of spatial locations.}

\item{fdr.proc}{Multiple hypothesis correction procedure, see
\code{multtest}.}
}
\value{
A data.frame with the names of the interactions, the value of the
chosen statistics, and the corresponding Q-value.
}
\description{
Compute the statistical association of L-R interaction score spatial
distributions with tissue area labels.
Not limited to BulkSignalR gene signature scores.
}
\details{
In case the
nonparametric Kruskal-Wallis test is chosen, additional columns are provided
testing each label for significantly larger scores (Kruskal-Wallis is global
and only says whether one or several labels show a bias). Individual
labels are tested with Wilcoxon and two columns are added *per* label,
one for the statistics and one for a Bonferroni-corrected P-value over
all the labels.

In case the parametric model is chosen, a linear model followed by ANOVA
are used for the global association test. Individual labels are tested
with T-tests (Bonferroni-corrected).
}
\examples{
print('spatialAssociation')
if(FALSE){

spatialAssociation(scores, areas)

}
}
