% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatialAssociationPlot}
\alias{spatialAssociationPlot}
\title{Heatmap plot of association of scores with area labels}
\usage{
spatialAssociationPlot(
  associations,
  qval.thres = 0.01,
  absval.thres = 0,
  colors = NULL
)
}
\arguments{
\item{associations}{A statistical association data.frame generated
by the function \code{spatialAssociation}.}

\item{qval.thres}{The maximum Q-value to consider in the plot (a
L-R interaction must associate with one label at least with a Q-value
smaller or equal to this threshold).}

\item{absval.thres}{The minimum value to consider in the plot (a
L-R interaction must associate with one label at least with an absolute
value larger or equal to this threshold).}

\item{colors}{A function returning a color for a given value such as
generated by \code{circlize::colorRamp2}.}
}
\value{
ComplexHeatmap::Heatmap object
}
\description{
Plot a heatmap featuring Q-values 
or values of statistical association between
L-R interaction score spatial distributions and tissue area labels.
}
\details{
Display a heatmap linking L-R interactions to labels.
}
\examples{
data(bsrdm.spa)
data(bsrinf.spa)

library(STexampleData)
spe <- Visium_humanDLPFC()

#set.seed(123)
#idx <- sample(ncol(speSubset), 10)
name.idx <- c("10x32","3x47","4x50",
"17x111","5x59","0x20","8x100",
"8x108","14x30","11x39")                                       
idx <- c(415,463, 179, 195, 118, 299, 229, 244, 14, 374) 

speSubset <- spe[, colData(spe)$ground_truth\%in\%c("Layer1","Layer2")]

speSubset <- speSubset[, idx]


colData(speSubset)$idSpatial <- paste(colData(speSubset)[[4]],
                colData(speSubset)[[5]],sep = "x")


annotation <- colData(speSubset)

thres <- 0.01
bsrinf.red <- reduceToBestPathway(bsrinf.spa)
s.red  <- getLRGeneSignatures(bsrinf.red, qval.thres=thres)
scores.red <- scoreLRGeneSignatures(bsrdm.spa,s.red)

# statistical association with tissue areas based on correlations
# For display purpose, we only use a subset here

assoc.bsr.corr <- spatialAssociation(scores.red[c(1:10), ],
areas = annotation, label.col = "ground_truth",test = "Spearman")

spatialAssociationPlot(assoc.bsr.corr)
}
