% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatialIndexPlot}
\alias{spatialIndexPlot}
\title{Generate a visual index of spatial score distributions}
\usage{
spatialIndexPlot(
  scores,
  areas,
  out.file,
  image.raster = NULL,
  dot.size = 0.25,
  ratio = 1.25,
  base.v = 2.5,
  base.h = 3
)
}
\arguments{
\item{scores}{A matrix of scores, one L-R interaction per row and
spatial locations in the columns. This matrix is typically obtained
from BulkSignalR functions \code{scoreLRGeneSignatures} or \code{scScoring}.}

\item{areas}{A data.frame containing at least the x and y
coordinates of the locations, the unique IDs of spatial locations, and
a tissue label column.}

\item{out.file}{File name for the output PDF.}

\item{image.raster}{Raster object image to plot raw tissue image as
reference.}

\item{dot.size}{Dot size.}

\item{ratio}{the vertical/horizontal ratio.}

\item{base.v}{Height of each plot.}

\item{base.h}{Width of each plot.}
}
\description{
Generate an index made of series of small individual spatial score plots
in a PDF. Not limited to BulkSignalR gene signature scores.
}
\details{
A PDF file is created that contains the index.
}
\examples{
print('spatialIndexPlot')
if(FALSE){

spatialIndexPlot(scores, areas, out.file)

}
}
