% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{spatialPlot}
\alias{spatialPlot}
\title{Print spatial plots}
\usage{
spatialPlot(
  v,
  areas,
  inter.name,
  rev.y = FALSE,
  rev.x = FALSE,
  ref.plot = FALSE,
  image.raster = NULL,
  x.col = "array_col",
  y.col = "array_row",
  label.col = "label",
  idSpatial.col = "idSpatial",
  cut.p = 0.01,
  low.color = "royalblue3",
  mid.color = "white",
  hi.color = "orange",
  title.fs = 12,
  legend.fs = 10,
  axis.fs = 10,
  label.fs = 12
)
}
\arguments{
\item{v}{Named vector with id of spots and signal.}

\item{areas}{Dataframe with coordinates and signal.
It should contains columns for id,
and a column for the signal you want to visualize.}

\item{inter.name}{Interaction name used as title.}

\item{rev.y}{Reverse y-axis.}

\item{rev.x}{Reverse x-axis.}

\item{ref.plot}{Logical to know if we plot reference object.}

\item{image.raster}{Raster object image.}

\item{x.col}{x coordinate}

\item{y.col}{y coordinate}

\item{label.col}{Labels}

\item{idSpatial.col}{Spot ids}

\item{cut.p}{Remove cut.p of extreme values}

\item{low.color}{"Low" color for gradient.}

\item{mid.color}{"Medium" color for gradient.}

\item{hi.color}{"High" color for gradient.}

\item{title.fs}{Fontsize for title.}

\item{legend.fs}{Fontsize for legend}

\item{axis.fs}{Fontsize for axis.}

\item{label.fs}{Fontsize for label.}
}
\value{
A ggplot object
}
\description{
By default, user will print signal on a 2D plot.
}
\details{
User can 
specifiy to plot on the left side,
a reference plot.
Reference plot is the raw image or
image with different colors for
all specified regions in \code{areas}.
}
\examples{
print('spatialPlot')
if(FALSE){
img <- png::readPNG(path.to.file)

spatialPlot(v,
    areas,
    inter.name="L->R",
    image.raster=img)
}
}
