% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R
\name{srcCmpName<-,BSRInferenceComp-method}
\alias{srcCmpName<-,BSRInferenceComp-method}
\title{Source comparison name setter (internal use only)}
\usage{
\S4method{srcCmpName}{BSRInferenceComp}(x) <- value
}
\arguments{
\item{x}{BSRInferenceComp object}

\item{value}{value to be set for bsrinf}
}
\value{
returns \code{NULL}
}
\description{
Source comparison name setter (internal use only)
}
\keyword{internal}
