% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{stats<-,BSRClusterComp-method}
\alias{stats<-,BSRClusterComp-method}
\title{Cluster comparison statistics setter (internal use only)}
\usage{
\S4method{stats}{BSRClusterComp}(x) <- value
}
\arguments{
\item{x}{object BSRClusterComp}

\item{value}{value to be set for BSRClusterComp}
}
\description{
Cluster comparison statistics setter (internal use only)
}
\keyword{internal}
