% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{tgCorr,BSRInference-method}
\alias{tgCorr,BSRInference-method}
\title{Target gene correlations accessor}
\usage{
\S4method{tgCorr}{BSRInference}(x)
}
\description{
Target gene correlations accessor
}
