% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{tgCorr<-,BSRInference-method}
\alias{tgCorr<-,BSRInference-method}
\title{Target gene correlations setter (internal use only)}
\usage{
\S4method{tgCorr}{BSRInference}(x) <- value
}
\arguments{
\item{x}{BSRInference object}

\item{value}{value to be set for bsrinf}
}
\description{
Target gene correlations setter (internal use only)
}
\keyword{internal}
