% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R, R/BSRSignature-class.R
\name{tgCorr}
\alias{tgCorr}
\alias{tgCorr,BSRInference-method}
\alias{tgCorr,BSRSignature-method}
\title{Target gene correlations accessor}
\usage{
\S4method{tgCorr}{BSRInference}(x)

\S4method{tgCorr}{BSRSignature}(x)
}
\arguments{
\item{x}{BSRSignature}
}
\description{
Target gene correlations accessor

Target gene correlations accessor
}
