% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{tGenes<-,BSRInference-method}
\alias{tGenes<-,BSRInference-method}
\title{Target genes setter (internal use only)}
\usage{
\S4method{tGenes}{BSRInference}(x) <- value
}
\description{
Target genes setter (internal use only)
}
