% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R, R/BSRSignature-class.R
\name{tGenes}
\alias{tGenes}
\alias{tGenes,BSRInference-method}
\alias{tGenes,BSRSignature-method}
\title{Target genes accessor}
\usage{
\S4method{tGenes}{BSRInference}(x)

\S4method{tGenes}{BSRSignature}(x)
}
\arguments{
\item{x}{BSRSignature}
}
\description{
Target genes accessor

Target genes accessor
}
