% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R
\name{tgExpr<-,BSRInferenceComp-method}
\alias{tgExpr<-,BSRInferenceComp-method}
\title{Target gene expression setter (internal use only)}
\usage{
\S4method{tgExpr}{BSRInferenceComp}(x) <- value
}
\arguments{
\item{x}{BSRInferenceComp object}

\item{value}{value to be set for bsrinf}
}
\value{
returns \code{NULL}
}
\description{
Target gene expression setter (internal use only)
}
\keyword{internal}
