% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{tgGenes<-,BSRInference-method}
\alias{tgGenes<-,BSRInference-method}
\title{Target genes setter (internal use only)}
\usage{
\S4method{tgGenes}{BSRInference}(x) <- value
}
\arguments{
\item{x}{BSRInference object}

\item{value}{value to be set BSRInference}
}
\value{
returns \code{NULL}
}
\description{
Target genes setter (internal use only)
}
\keyword{internal}
