% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R, R/BSRSignature-class.R
\name{tgGenes}
\alias{tgGenes}
\alias{tgGenes,BSRInference-method}
\alias{tgGenes,BSRSignature-method}
\title{Target genes accessor}
\usage{
\S4method{tgGenes}{BSRInference}(x)

\S4method{tgGenes}{BSRSignature}(x)
}
\arguments{
\item{x}{BSRSignature}
}
\value{
tgGenes

tgGenes
}
\description{
Target genes accessor

Target genes accessor
}
\examples{
bsr.sig <- new("BSRSignature")
tgGenes(bsr.sig)
}
