% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R
\name{tgLogFC<-,BSRInferenceComp-method}
\alias{tgLogFC<-,BSRInferenceComp-method}
\title{Target gene logFC setter (internal use only)}
\usage{
\S4method{tgLogFC}{BSRInferenceComp}(x) <- value
}
\arguments{
\item{x}{BSRInferenceComp object}

\item{value}{value to be set for bsrinf}
}
\description{
Target gene logFC setter (internal use only)
}
\keyword{internal}
