% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tme.signatures}
\alias{tme.signatures}
\title{Tumor microenvironment gene signatures}
\format{
A data frame with 209 rows and 2 variables: \describe{
  \item{gene}{HUGO gene symbol} \item{signature}{cell population name} }
}
\source{
Becht & al., Genome Biol, 2016; Angelova et al., Genome Biol, 2015.
}
\usage{
data(tme.signatures)
}
\description{
A dataset containing gene signatures for some immune and stromal
  cell populations that are present in the microenvironment of a tumor.
}
\keyword{datasets}
