% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{alluvialPlot}
\alias{alluvialPlot}
\title{Alluvial plot}
\usage{
alluvialPlot(bsrinf, keywords, type = c("L", "R", "pw.id"), qval.thres = 0.01)
}
\arguments{
\item{bsrinf}{object bsrinf inference.}

\item{keywords}{vector of pathways.}

\item{type}{filter on Ligand, Receptor or pathway id.}

\item{qval.thres}{threshold over Q-value.}
}
\value{
NULL

This is a convenience function that relies on the \code{ggalluvial}
package to propose a simple way
of representing Ligands, Receptors
}
\description{
Representation of the links
between Ligands,Receptors and Pathways.
}
\examples{
data(bsrinf, package = "BulkSignalR")
alluvialPlot(bsrinf,
    keywords = c("LAMC1"),
    type = "L",
    qval.thres = 0.01)
}
