% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{alluvialPlot}
\alias{alluvialPlot}
\title{Alluvial plot}
\usage{
alluvialPlot(
  bsrinf,
  keywords,
  type = c("L", "R", "pw.id"),
  qval.thres = 0.01,
  format = c("pdf", "svg", "png"),
  path = "./",
  filename = NULL,
  width = 10,
  height = 10
)
}
\arguments{
\item{bsrinf}{object bsrinf inference.}

\item{keywords}{vector of pathways.}

\item{type}{filter on Ligand, Receptor or pathway id.}

\item{qval.thres}{threshold over Q-value.}

\item{format}{pdf / png / sg. By default, it will 
plot in pdf format.}

\item{path}{directory where to plot file.}

\item{filename}{file name, NULL by default (plot on screen).}

\item{width}{width of image in cm.}

\item{height}{height of image in cm.}
}
\value{
NULL

This is a convenience function that relies on the \code{ggalluvial}
package to propose a simple way
of representing Ligands, Receptors
}
\description{
Representation of the links
between Ligands,Receptors and Pathways.
}
\examples{
print('alluvialPlot')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
alluvialPlot(bsrinf,
             keywords = c("COL4A1"),
             type = "L",
             qval.thres = 0.001,
             path = "./",
             filename = "sdc_alluvial", 
             width  = 16, 
             height = 12
             )
}
