% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bodyMap.mouse}
\alias{bodyMap.mouse}
\title{Mouse transcriptomes across tissues}
\format{
A data frame with 24543 rows and 8 variables.
}
\source{
Bin Li & al., Scientific Reports, 2017;
}
\usage{
data(bodyMap.mouse)
}
\description{
A dataset containing rpkm values of  brain and liver.
}
\keyword{datasets}
