% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\docType{class}
\name{BSRClusterComp-class}
\alias{BSRClusterComp-class}
\title{BulkSignalR Cluster Comparison Object}
\description{
An S4 class to represent the comparison of two clusters of samples to
infer LR interactions based on the resulting P-values and
log-fold-changes (logFC).
}
\section{Slots}{

\describe{
\item{\code{colA}}{Column indices for the samples in cluster A.}

\item{\code{colB}}{Column indices for the samples in cluster B.}

\item{\code{stats}}{Comparison statistics A versus B as a data.frame and
containing at least two columns named 'pval' and 'logFC'.}
}}

\examples{
bsrdm <- new("BSRDataModel", ncounts=matrix(1.5, nrow=2, ncol=4,
             dimnames=list(c("A","B"), c("E","F","G","H"))),
             log.transformed=TRUE, normalization="TC")
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:2)
colB <- as.integer(3:4)
n <- nrow(ncounts(bsrdm.comp))
edger.stats <- data.frame(pval=runif(n), logFC=rnorm(n, 0, 2))
rownames(edger.stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, edger.stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "my_comparison")

}
