% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModelComp-class.R
\docType{class}
\name{BSRDataModelComp-class}
\alias{BSRDataModelComp-class}
\title{BulkSignalR Data Model Compare Object}
\description{
An S4 class to represent the expression data used for inferring
ligand-receptor interactions based on sample cluster comparisons.
}
\section{Slots}{

\describe{
\item{\code{comp}}{A named list of BSRClusterComp objects, one per
comparison.}
}}

\examples{
# prepare data
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:5)
colB <- as.integer(8:15)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(pval=runif(n), logFC=rnorm(n, 0, 2))
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison
bsrinf <- initialInference(bsrdm.comp, max.pval=1,"random.example")

}
