% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsrinf.comp}
\alias{bsrinf.comp}
\title{A skinny BSR-InferenceComp object related to sdc.}
\format{
An example of an BSR-InferenceComp object
}
\usage{
data(bsrinf.comp)
}
\description{
See Vignette BulkSignalR-Differential.
}
\keyword{datasets}
