% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R
\docType{class}
\name{BSRInferenceComp-class}
\alias{BSRInferenceComp-class}
\title{BulkSignalR cluster comparison-based inference object}
\description{
An S4 class to represent ligand-receptor interactions inferred from
a comparison of two clusters of samples. This class inherits from
BSRInference.
}
\details{
This class is contains inferred LR interactions along with
their statistical significance. Data representation supports subsequent
reductions to pathways, etc. See reduction functions
\code{"\link[=BSRInferenceComp-class]{reduceToBestPathway}"},
\code{"\link[=BSRInferenceComp-class]{reduceToLigand}"},
\code{"\link[=BSRInferenceComp-class]{reduceToReceptor}"} and
\code{"\link[=BSRInferenceComp-class]{reduceToPathway}"}.
}
\section{Slots}{

\describe{
\item{\code{cmp.name}}{The name of the BSRClusterComp object in a BSRDataModelComp
object comp list.}

\item{\code{tg.pval}}{A list of target gene P-values, one
entry per interaction}

\item{\code{tg.logFC}}{A list of target gene logFC, one
entry per interaction}
}}

\examples{
print('BSRInferenceComp class')

# prepare data
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:5)
colB <- as.integer(8:15)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(pval=runif(n), logFC=rnorm(n, 0, 2))
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison
bsrinf <- initialInference(bsrdm.comp,max.pval=1,"random.example")

}
