% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cacheClear}
\alias{cacheClear}
\title{Delete cache content.}
\usage{
cacheClear(dir = c("resources"))
}
\arguments{
\item{dir}{Directory to remove. Can be only 'resources'.}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Delete the content of cache directory.
}
\examples{
cacheClear(dir="resources")
# need to recreate database in order to run examples well
createResources(verbose=TRUE)
}
