% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cacheClear}
\alias{cacheClear}
\title{Delete cache content.}
\usage{
cacheClear(dir = c("both", "resources", "database"))
}
\arguments{
\item{dir}{Directory to remove. Can be only 'resources' or 'database'.}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Delete the content of cache directory.
}
\examples{
print("cacheClear")
}
