% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cacheInfo}
\alias{cacheInfo}
\title{Get cache content information.}
\usage{
cacheInfo(dir = c("resources"))
}
\arguments{
\item{dir}{Directory to remove in order to clean the cache.
Can be only 'resources'}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Get cache content information for a specific cache directory.
}
\examples{
cacheInfo()
}
