% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cacheInfo}
\alias{cacheInfo}
\title{Get cache content informations..}
\usage{
cacheInfo(dir = c("both", "resources", "database"))
}
\arguments{
\item{dir}{Directory to remove in order to clean the cache.
Can be only 'resources', 'database' or 'both'.}
}
\description{
Get cache content informations for specific cache dir.
}
\examples{
print("cacheInfo")
}
