% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cacheVersion}
\alias{cacheVersion}
\title{Check whether remote resource files have been changed.}
\usage{
cacheVersion(dir = c("resources"))
}
\arguments{
\item{dir}{Directory for which you want to check Version.
Can be only 'resources'.}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Check to see whether some resource
has been updated.
}
\examples{
cacheVersion()
}
