% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cacheVersion}
\alias{cacheVersion}
\title{Check remote files ressources are changed.}
\usage{
cacheVersion(dir = c("both", "resources", "database"))
}
\arguments{
\item{dir}{Directory for which you want to check Version.
Can be only 'resources', 'database' or 'both'.}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Check to see if some ressources has
has been updated.
}
\examples{
print("cacheVersion")
}
