% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeAssign.R
\name{cellTypeFrequency}
\alias{cellTypeFrequency}
\title{Cell type frequencies in relations to gene sets}
\usage{
cellTypeFrequency(rel, lr, min.n.genes = 1)
}
\arguments{
\item{rel}{The data.frame output by
\code{"\link[=BSRInference-class]{relateToGeneSet}"}.}

\item{lr}{The data.frame output by
\code{"\link{assignCellTypesToInteractions}"}.}

\item{min.n.genes}{Minimum number of genes in the gene set for
one (L,R,pathway) triple.}
}
\value{
A list of two slots: t for counting how many times each cell type
is involved; s for summing the weights of each involved cell type.
}
\description{
Count how many times and with which weights cell types were involved
in the (L,R,pathway) triples that targeted genes in a gene set.
}
\examples{
print('assignCellTypesToInteractions')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)

# microenvironment cell populations
data(immune.signatures, package="BulkSignalR")
immune.signatures <- immune.signatures[immune.signatures$signature \%in\%
                               c("B cells","Dentritic cells","Macrophages",
                                 "NK cells","T cells","T regulatory cells"),]
data("tme.signatures", package="BulkSignalR")
signatures <- rbind(immune.signatures,tme.signatures[
          tme.signatures$signature\%in\%c("Endothelial cells","Fibroblasts"),])
tme.scores <- scoreSignatures(bsrdm, signatures)

# assignment
lr2ct <- assignCellTypesToInteractions(bsrdm, bsrinf, tme.scores)

# relate to p-EMT (should be done in HNSCC normally, not in SDC)
data(p.EMT, package='BulkSignalR')
p.EMT <- p.EMT$gene
triggers <- relateToGeneSet(bsrinf, p.EMT)

# counts
cf <- cellTypeFrequency(triggers, lr2ct)
}
