% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{chordDiagramLR}
\alias{chordDiagramLR}
\title{Chord Diagram of LR interactions with correlations}
\usage{
chordDiagramLR(
  bsrinf,
  path = "./",
  filename = "chord",
  pw.id.filter = "R-RSA-17821",
  ligand = "L1",
  receptor = "R1",
  limit = 20,
  format = c("svg", "png", "pdf"),
  width = 4,
  height = 4
)
}
\arguments{
\item{bsrinf}{bsrinf object}

\item{path}{Path where to plot}

\item{filename}{Filename for the plot}

\item{pw.id.filter}{One Pathway ID accepted only to}

\item{ligand}{Ligand
for the LR pair that you want to 
highlight in the chord diagram.}

\item{receptor}{Receptor
for the LR pair that you want to highlight
in the chord diagram.}

\item{limit}{Number of interactions you can visualize.}

\item{format}{svg / png / pdf. By default means it will 
plot in a svg format.}

\item{width}{width of image}

\item{height}{height of image}
}
\value{
Circos Plot on Disk
}
\description{
By default, chord diagrams will be plot on disk.
}
\examples{
print('chordDiagramLR')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
chordDiagramLR (bsrinf,
                 path="./",
                 filename="sdc_chord",
                 pw.id.filter="R-HSA-202733",
                 ligand="COL18A1",
                 receptor="ITGA3",
                 limit=20,
                 width=5, 
                 height=4.5
   )
}
