% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R,
%   R/BSRSignatureComp-class.R
\name{cmpName}
\alias{cmpName}
\alias{cmpName,BSRInferenceComp-method}
\alias{cmpName,BSRSignatureComp-method}
\title{Comparison name accessor}
\usage{
\S4method{cmpName}{BSRInferenceComp}(x)

\S4method{cmpName}{BSRSignatureComp}(x)
}
\arguments{
\item{x}{BSRSignature}
}
\description{
Comparison name accessor

Target genes accessor
}
