% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModelComp-class.R
\name{coerce}
\alias{coerce}
\title{Convert BSRDataModel toBSRDataModelComp}
\arguments{
\item{from}{BSRDataModel object}
}
\value{
A BSRDataModelComp object
}
\description{
Convert BSRDataModel toBSRDataModelComp
}
\examples{
bsrdm <- new("BSRDataModel")
bsrdm.comp <- as(bsrdm, "BSRDataModelComp")

}
