% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{colB}
\alias{colB}
\alias{colB,BSRClusterComp-method}
\title{Cluster B columns accessor}
\usage{
\S4method{colB}{BSRClusterComp}(x)
}
\arguments{
\item{x}{object BSRClusterComp}
}
\description{
Cluster B columns accessor
}
