% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModelComp-class.R
\name{comp}
\alias{comp}
\alias{comp,BSRDataModelComp-method}
\title{Comparisons list accessor}
\usage{
\S4method{comp}{BSRDataModelComp}(x)
}
\arguments{
\item{x}{object BSRDataModelComp}
}
\description{
Comparisons list accessor
}
