% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{coordsFlip}
\alias{coordsFlip}
\title{(Counter-)clockwise 90 degree rotation}
\usage{
coordsFlip(rasterImage, degrees = 90)
}
\arguments{
\item{rasterImage}{Raster image.}

\item{degrees}{Rotation angle in degrees.}
}
\value{
A rotated raster image.
}
\description{
(Counter-)clockwise 90 degree rotation
}
\examples{
library(STexampleData)
spe <- Visium_humanDLPFC()
my.image.as.raster <- SpatialExperiment::imgRaster(spe, 
sample_id = imgData(spe)$sample_id[1], image_id = "lowres")
my.image.rotated <- coordsFlip(my.image.as.raster, degrees = 90)
}
