% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{coordsFlip}
\alias{coordsFlip}
\title{(Counter-)clockwise 90 degree rotation}
\usage{
coordsFlip(rasterImage, degrees = 90)
}
\arguments{
\item{rasterImage}{Raster image.}

\item{degrees}{Rotation angle in degrees.}
}
\value{
A rotated raster image.
}
\description{
(Counter-)clockwise 90 degree rotation
}
\examples{
print('coordsFlip')
if(FALSE){
img <- png::readPNG(path.to.file)
my.image.as.raster <- grDevices::as.raster(path.to.file)  
my.image.rotated <- coordsFlip(my.image.as.raster,degrees=90)

 }
}
