% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{createDatabase}
\alias{createDatabase}
\title{Fetch the database from internet.}
\usage{
createDatabase(onRequest = TRUE, verbose = FALSE)
}
\arguments{
\item{onRequest}{logical True if you force
download again. This will overwrite
pre-existing database. Default is True.}

\item{verbose}{Logical TRUE/FALSE}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Fetch LR database from remote location.
}
\examples{
print("createDatabase")
if(FALSE){
 createDatabase()
}
}
