% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{createResources}
\alias{createResources}
\title{Create all resources}
\usage{
createResources(onRequest = TRUE, verbose = FALSE)
}
\arguments{
\item{onRequest}{logical TRUE if you want to force
downloading again. This will overwrite the
pre-existing local database. Default is TRUE.}

\item{verbose}{Default is FALSE}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Create a cache for all resources (pathways, lrdb & network)
downloaded from the web when the library is first loaded.
This functionality is handled with BiocFileCache.
}
\examples{
createResources(onRequest=FALSE)
}
