% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.buildPermutatedCountMatrix}
\alias{.buildPermutatedCountMatrix}
\title{Internal function to generate a randomized expression matrix}
\usage{
.buildPermutatedCountMatrix(ncounts, pind)
}
\arguments{
\item{ncounts}{A matrix or table of normalized read counts.}

\item{pind}{Permutation indices such as returned by \code{.buildPermutationIndices}.}
}
\value{
\code{ncount} with shuffled row names (gene symbols). Shuffling is performed within rows with comparable average expression.
}
\description{
Internal function to generate a randomized expression matrix
}
