% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cacheAdd}
\alias{.cacheAdd}
\title{Add cache for resources.}
\usage{
.cacheAdd(fpath, cacheDir, resourceName, verbose = FALSE, download = TRUE)
}
\arguments{
\item{fpath}{Path to file on the web or local system.}

\item{cacheDir}{Absolute path to cache directory.}

\item{resourceName}{Ressource name.}

\item{verbose}{Default FALSE}

\item{download}{Logical(TRUE/FALSE) Default TRUE for download.}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Add cache for resources (pathways, lrdb, or network)
downloaded from the web or local database.
This part is handled with BiocFileCache.
}
\keyword{internal}
