% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrepare.R
\name{.checkInteroperabilityForCounts}
\alias{.checkInteroperabilityForCounts}
\title{Internal function to check and extract
counts matrix if a more complex object
is given as parameter.}
\usage{
.checkInteroperabilityForCounts(
  counts,
  symbol.col,
  x.col,
  y.col,
  barcodeID.col
)
}
\arguments{
\item{counts}{A table or matrix of read counts.
It can also be a SummarizedExperiment or SpatialExperiment
object from which counts matrix are extracted.
See \code{\link{BSRDataModel}}.}

\item{symbol.col}{The index of the column containing the gene symbols in case
those are not the row names of \code{counts} already. In a
SpatialExperiment object, the index in the dafaframe returned by rowData().}

\item{x.col}{In a SpatialExperiment object, the index of the column
containing x coordinates in the dafaframe returned by rowData(), usually 
named array_row}

\item{y.col}{In a SpatialExperiment object, the index of the column
containing y coordinates in the dafaframe returned by rowData(), usually 
named array_col}

\item{barcodeID.col}{In a SpatialExperiment object, the index of the column
containing barcodeID in the dafaframe returned by colData(), usually named
barcode_id}
}
\value{
A matrix of counts
}
\description{
Internal function to check and extract
counts matrix if a more complex object
is given as parameter.
}
\keyword{internal}
