% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.customheatmap}
\alias{.customheatmap}
\title{Heatmap function for gene expression of signature}
\usage{
.customheatmap(
  counts,
  h.width = 5,
  h.height = 10,
  fontsize = 6,
  scoring = c(-1.5, 0, 4, 5, 6.1, 0.3),
  cols.scoring,
  hcl.palette = "Blues 3",
  show_column_names = FALSE
)
}
\arguments{
\item{counts}{Matrice of counts exported from a}

\item{h.width}{Heatmap  width in cm.}

\item{h.height}{Heatmap  height in cm.}

\item{fontsize}{Fontsize.}

\item{scoring}{Vector of scored sample for a 
a previously choosen pathway.}

\item{cols.scoring}{Fixed colorRamp2 object.}

\item{hcl.palette}{Palette from 
HCL colormaps supported by ComplexHeatmap.}

\item{show_column_names}{Add column names on heatmap.}
}
\value{
ComplexHeatmap object. 

This is a convenience function that relies 
on the \code{ComplexHeatmap}
package to propose a simple way
of representing expression of genes involved in a specific
pathway.
}
\description{
Generate one heatmap re-used by 
by  
\code{"\link[=BSRDataModel-class]{scoreLRGeneSignatures}"}
}
\keyword{internal}
