% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{.edgesLRIntracell}
\alias{.edgesLRIntracell}
\title{Internal function to generate a ligand-receptor-downstream signaling network}
\usage{
.edgesLRIntracell(pairs, pw, id.col, gene.col, min.cor = 0.3, signed = FALSE)
}
\arguments{
\item{pairs}{A ligand-receptor table such as output by \code{pValuesLR} and \code{naiveBayesLR}.}

\item{pw}{A table defining the reference pathways.}

\item{id.col}{Column index or name in \code{pw} for the pathway IDs.}

\item{gene.col}{Column index or name in \code{pw} for the gene symbols.}

\item{min.cor}{Minimum correlation required for the target genes.}

\item{signed}{A logical indicating whether \code{min.cor} is imposed to correlation absolute values (\code{FALSE}) or not (\code{TRUE}).}
}
\value{
An \code{igraph} object featuring the ligand-receptor-downstream signaling network. Default colors and node sizes are assigned.
}
\description{
Internal function to generate a ligand-receptor-downstream signaling network
}
