% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.formatPathwaysFromGmt}
\alias{.formatPathwaysFromGmt}
\title{Transform gmt file to dataframe}
\usage{
.formatPathwaysFromGmt(file, resourceName = NULL)
}
\arguments{
\item{file}{Path to GMT file}

\item{resourceName}{Two options "GO-BP" or "REACTOME"}
}
\value{
Dataframe with pathwayID, geneName and pathwayName
}
\description{
We note discrepancy between format available 
over internet.
}
\details{
Here we consider a valid gmt file format defined
on each lines as follows :  
First is Pathway name, 
Then comes the ID,
Finally you will find genes symbols
according to the pathway defined on the line.

You can find an example here.
- For Reactome. (Directly from their website)
\url{https://reactome.org/download/current/ReactomePathways.gmt.zip}
Note that you need to unzip the file to read the content.
The code is inspired from read.gmt function
from the gsa R package.
}
