% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{.geneNameConversion}
\alias{.geneNameConversion}
\title{Convert gene symbol to another organism}
\usage{
.geneNameConversion(genes, conversion.dict)
}
\arguments{
\item{genes}{genes you want to convert}

\item{conversion.dict}{A data frame containing
gene names for source species and Homo Sapiens.}
}
\value{
Depend type of input genes
LRinter return a vector of genes
tGenes receptors ligands : return list of list of genes
}
\description{
Convert gene symbol to another organism
based on a dictionary with human and ortholog species.
}
\keyword{internal}
