% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTraining.R
\name{.getEmpiricalNullCorrLR}
\alias{.getEmpiricalNullCorrLR}
\title{Sampling of ligand-receptor correlation null distribution}
\usage{
.getEmpiricalNullCorrLR(obj)
}
\arguments{
\item{obj}{A BSRDatamodel without learned paramaters.}
}
\value{
A list of \code{n.rand.LR} tables such as output by
  \code{\link{.getCorrelatedLR}}. Each table is computed from a randomized
  expression matrix (randomized \code{ncounts}).
}
\description{
Perform a ligand-receptor Spearman correlation analysis based
on randomized expression data.
}
\details{
A large number of correlations is reported in each randomized
  matrix. Therefore, \code{n.rand.LR} 
  should be given a modest value to avoid unnecessarily long
  computations.

  See \code{\link{.getCorrelatedLR}} for more details about the parameters.
}
\keyword{internal}
