% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrepare.R
\name{.testRemoteServer}
\alias{.testRemoteServer}
\title{Check whether the reference DB server is up}
\usage{
.testRemoteServer()
}
\value{
Returns `NULL`, invisibly.
}
\description{
Check whether the reference DB server is up
}
