% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getComplexes}
\alias{getComplexes}
\title{Retrieve LR complexes}
\usage{
getComplexes(idRelease = NULL)
}
\arguments{
\item{idRelease}{integer id version Release
Default is NULL so last version is selected.}
}
\value{
Returns dataframe Complexex, invisibly.
}
\description{
Fetch LR complexes from database and
and return a dataframe
}
\examples{
print("getComplexes")
getComplexes(idRelease=1)
}
