% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getInteractions}
\alias{getInteractions}
\title{Retrieve LR interactions.}
\usage{
getInteractions(idRelease = NULL)
}
\arguments{
\item{idRelease}{integer id version Release
Default is NULL so last version is selected.}
}
\description{
Fetch LR interactions from database and
and return a dataframe
}
\examples{
print("getInteractions")
getInteractions()
}
