% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{getLRIntracellNetwork}
\alias{getLRIntracellNetwork}
\title{Generate a ligand-receptor-downstream signaling network}
\usage{
getLRIntracellNetwork(
  bsrinf,
  pval.thres = NULL,
  qval.thres = NULL,
  min.cor = 0.3,
  restrict.pw = NULL,
  node.size = 5
)
}
\arguments{
\item{bsrinf}{A BSRInference object.}

\item{pval.thres}{P-value threshold.}

\item{qval.thres}{Q-value threshold.}

\item{min.cor}{Minimum correlation required for the target genes.}

\item{restrict.pw}{A vector of pathway IDs to which receptor downstream
signaling is restricted.}

\item{node.size}{Default node size in the network.}
}
\value{
An \code{igraph} object featuring the ligand-receptor-downstream
signaling network. Default colors and node sizes are assigned,
which can be changed afterwards if necessary.

The target genes to which the \code{min.cor} correlation is imposed are
those listed in \code{tGenes(bsrinf)}, correlations are in
\code{tgCorr(bsrinf)}.
The construction of shortest paths from the receptors to those selected
targets adds other genes, which were either some targets with too low
correlation or genes along the shortest paths to reach the selected targets.
}
\description{
Generate a ligand-receptor network from a BSRInference object and add
the shortest paths from the receptors to correlated
target genes following Reactome and KEGG pathways.
}
\examples{
\dontrun{
}
}
