% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{getLRIntracellNetwork}
\alias{getLRIntracellNetwork}
\title{Generate a ligand-receptor-downstream signaling network}
\usage{
getLRIntracellNetwork(
  pairs,
  LLR.thres = NULL,
  pval.thres = NULL,
  qval.thres = NULL,
  min.cor = 0.3,
  signed = FALSE,
  restrict.pw = NULL,
  node.size = 5
)
}
\arguments{
\item{pairs}{A ligand-receptor table such as output by \code{pValuesLR} and \code{naiveBayesLR}.}

\item{LLR.thres}{Log-likelihood threshold.}

\item{pval.thres}{P-value threshold.}

\item{qval.thres}{Q-value threshold.}

\item{min.cor}{Minimum correlation required for the target genes.}

\item{signed}{A logical indicating whether \code{min.cor} is imposed to correlation absolute values (\code{FALSE}) or not (\code{TRUE}).}

\item{restrict.pw}{A vector of pathway IDs to which receptor downstream signaling is restricted.}

\item{node.size}{Default node size in the network.}
}
\value{
An \code{igraph} object featuring the ligand-receptor-downstream signaling network. Default colors and node sizes are assigned,
which can be changed afterwards if necessary.

The target genes to which the \code{min.cor} correlation is imposed are those listed in \code{pairs$target.genes}, correlations are in \code{pairs$all.cor}.
The construction of shortest paths from the receptors to those selected targets adds other genes, which were either some targets with too little correlation
or genes along the shortest paths to reach the selected targets.
}
\description{
Generate a ligand-receptor network from a ligand-receptor table and add the shortest paths from the receptor to correlated
target genes following Reactome and KEGG pathways.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pp <- pValuesLR(ds.LR,ds$param)
gLRintra <- getLRIntracellNetwork(pp,qval.thres=0.01)
plot(gLRintra)
write.graph(gLRintra,file="SDC-LR-intracellular-network.graphml",format="graphml")
}
}
