% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{getResource}
\alias{getResource}
\title{Get ressource from the cache.}
\usage{
getResource(resourceName = NULL)
}
\arguments{
\item{resourceName}{Ressource name.}
}
\description{
Get  resources (pathways, or PathwayCommons network 
from \url{https://www.pathwaycommons.org/})
stored in the cache.
}
\examples{
reactome <-  getResource(resourceName="Reactome")
}
