% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R,
%   R/BSRDataModelComp-class.R
\name{initialInference}
\alias{initialInference}
\alias{initialInference,BSRDataModel-method}
\alias{initialInference,BSRDataModelComp-method}
\title{Inference of ligand-receptor interactions}
\usage{
\S4method{initialInference}{BSRDataModel}(
  obj,
  rank.p = 0.55,
  min.cor = 0.25,
  restrict.genes = NULL,
  reference = c("REACTOME-GOBP", "REACTOME", "GOBP"),
  max.pw.size = NULL,
  min.pw.size = NULL,
  min.positive = NULL,
  restrict.pw = NULL,
  with.complex = NULL,
  fdr.proc = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BY", "ABH",
    "TSBH")
)

\S4method{initialInference}{BSRDataModelComp}(
  obj,
  cmp.name,
  rank.p = 0.55,
  max.pval = 0.01,
  min.logFC = 1,
  neg.receptors = FALSE,
  restrict.genes = NULL,
  reference = c("REACTOME-GOBP", "REACTOME", "GOBP"),
  max.pw.size = 200,
  min.pw.size = 5,
  min.positive = 4,
  restrict.pw = NULL,
  with.complex = TRUE,
  fdr.proc = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BY", "ABH",
    "TSBH")
)
}
\arguments{
\item{obj}{A BSRDataModelComp object.}

\item{rank.p}{A number between 0 and 1 defining the rank of the last
considered target genes.}

\item{min.cor}{The minimum Spearman correlation required between
the ligand and the receptor.}

\item{restrict.genes}{A list of gene symbols that restricts ligands and
receptors.}

\item{reference}{Which pathway reference should be used ("REACTOME"
for Reactome, "GOBP" for GO Biological Process,
or "REACTOME-GOBP" for both).}

\item{max.pw.size}{Maximum pathway size to consider from the pathway
reference.}

\item{min.pw.size}{Minimum pathway size to consider from the pathway
reference.}

\item{min.positive}{Minimum number of target genes to be found in a given
pathway.}

\item{restrict.pw}{A list of pathway IDs to restrict the application of
the function.}

\item{with.complex}{A logical indicating whether receptor co-complex
members should be included in the target genes.}

\item{fdr.proc}{The procedure for adjusting P-values according to
\code{\link[multtest]{mt.rawp2adjp}}.}

\item{cmp.name}{The name of the cluster comparison that should be used for
the inference.}

\item{max.pval}{The maximum P-value imposed to both the ligand
and the receptor.}

\item{min.logFC}{The minimum log2 fold-change allowed for
both the receptor and the ligand.}

\item{neg.receptors}{A logical indicating whether receptors are only
allowed to be upregulated (FALSE), or up- and downregulated (TRUE).}
}
\value{
A BSRInference object with initial inferences set.

A BSRInferenceComp object with initial inferences set.
}
\description{
Computes putative LR interactions along with their statistical confidence.
In this initial inference, all the relevant pathways are reported,
see reduction functions to reduce this list.

Computes putative LR interactions along with their statistical confidence.
The analysis is based on P-values associated to the regulation of
genes/proteins in a comparison of two clusters of samples.
In this initial inference, all the relevant pathways are reported,
see reduction functions to reduce this list.
}
\details{
Perform the initial ligand-receptor inference. Initial means that
no reduction is applied. All the (ligand, receptor, downstream pathway)
triples are reported, i.e., a given LR pair may appear multiple times
with different pathways downstream the receptor. Specific reduction
functions are available from the package to operate subsequent
simplifications based on the BSRInference object created by the initial
inference.

Parameters defining minimum/maximum pathway sizes, etc. are set to NULL
by default, meaning that their values will be taken from what was set
during the training of the statistical model with
\code{"\link[=BSRDataModel-class]{learnParameters}"}

To use different
values at the time of inference sounds like a bad idea, although this
could be used to explore without retraining the underlying model.
Retraining of the model with adjusted parameters is advised following
such an exploration.

Perform the initial ligand-receptor inference. Initial means that
no reduction is applied. All the (ligand, receptor, downstream pathway)
triples are reported, i.e., a given LR pair may appear multiple times
with different pathways downstream the receptor. Specific reduction
functions are available from the package to operate subsequent
simplifications based on the BSRInferenceComp object created by this method.

Here, ligand-receptor interactions are inferred based on gene or protein
regulation-associated P-values when comparing two clusters of samples. Since
a BSRDataModelComp object can contain several such comparisons, the name
of the comparison to use must be specified (parameter \code{cmp.name}).
}
\examples{
print('prepareDataset')
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

print("learnParameters")
bsrdm <- learnParameters(bsrdm)
bsrdm

print('perform inference')
bsrinf <- initialInference(bsrdm)
bsrinf
# prepare data
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:5)
colB <- as.integer(8:15)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(pval=runif(n), logFC=rnorm(n, 0, 2))
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison
bsrinf <- initialInference(bsrdm.comp,max.pval=1, "random.example")
}
