% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{logTransformed}
\alias{logTransformed}
\alias{logTransformed,BSRDataModel-method}
\title{log.transformed accessor}
\usage{
\S4method{logTransformed}{BSRDataModel}(x)
}
\arguments{
\item{x}{Object BRSDataModel}
}
\description{
log.transformed accessor
}
