% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{LRinter<-,BSRInference-method}
\alias{LRinter<-,BSRInference-method}
\title{LRinter setter (internal use only)}
\usage{
\S4method{LRinter}{BSRInference}(x) <- value
}
\arguments{
\item{x}{BSRInference object}

\item{value}{value to be set to BSRInference}
}
\description{
LRinter setter (internal use only)
}
\keyword{internal}
