% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R
\name{LRinterScore}
\alias{LRinterScore}
\alias{LRinterScore,BSRInferenceComp-method}
\title{Simplified LRinter accessor with focus on the LR-score}
\usage{
\S4method{LRinterScore}{BSRInferenceComp}(x)
}
\arguments{
\item{x}{BSRInferenceComp object}
}
\value{
LRinterScore
}
\description{
Simplified LRinter accessor with focus on the LR-score
}
\examples{
data(bsrinf.comp, package = "BulkSignalR")
LRinterScore(bsrinf.comp)[5,]
}
