% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModelComp-class.R
\name{mu}
\alias{mu}
\alias{mu,BSRDataModelComp-method}
\title{Mu accessor}
\usage{
\S4method{mu}{BSRDataModelComp}(x)
}
\arguments{
\item{x}{object BSRDataModelComp}
}
\value{
mu
}
\description{
Mu accessor
}
\examples{
bsrdm.comp <- new("BSRDataModelComp")
mu(bsrdm.comp)
}
