% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{ncounts}
\alias{ncounts}
\alias{ncounts,BSRDataModel-method}
\title{Normalized count matrix accessor}
\usage{
\S4method{ncounts}{BSRDataModel}(x)
}
\arguments{
\item{x}{object BSRDataModel}
}
\description{
Normalized count matrix accessor
}
