% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{normalization}
\alias{normalization}
\alias{normalization,BSRDataModel-method}
\title{Normalization accessor}
\usage{
\S4method{normalization}{BSRDataModel}(x)
}
\arguments{
\item{x}{object BSRDatamModel}
}
\value{
normalization
}
\description{
Normalization accessor
}
\examples{
bsrdm <- new("BSRDataModel",
    ncounts = matrix(1.5,
        nrow = 2, ncol = 2,
        dimnames = list(c("A", "B"), c("C", "D"))
    ),
    log.transformed = TRUE,
    normalization = "TC"
)
normalization(bsrdm)
}
