% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{param}
\alias{param}
\alias{param,BSRDataModel-method}
\title{Model parameter accessor}
\usage{
\S4method{param}{BSRDataModel}(x)
}
\arguments{
\item{x}{BSRDataModel oject}
}
\description{
Model parameter accessor
}
