% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{parameters}
\alias{parameters}
\alias{parameters,BSRDataModel-method}
\title{Model parameter accessor}
\usage{
\S4method{parameters}{BSRDataModel}(x)
}
\arguments{
\item{x}{BSRDataModel oject}
}
\value{
param
}
\description{
Model parameter accessor
}
\examples{
bsrdm <- new("BSRDataModel",
    ncounts = matrix(1.5,
        nrow = 2, ncol = 2,
        dimnames = list(c("A", "B"), c("C", "D"))
    ),
    log.transformed = TRUE,
    normalization = "TC"
)
parameters(bsrdm)
}
