% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRSignature-class.R
\name{pathways}
\alias{pathways}
\alias{pathways,BSRSignature-method}
\title{pathways accessor}
\usage{
\S4method{pathways}{BSRSignature}(x)
}
\arguments{
\item{x}{BSRSignature}
}
\description{
pathways accessor
}
