% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rasterizeFromFile}
\alias{rasterizeFromFile}
\title{Rasterize raw image from filepath}
\usage{
rasterizeFromFile(path.to.file)
}
\arguments{
\item{path.to.file}{Path to an image file.}
}
\value{
A raster image object.
}
\description{
Rasterize raw image from filepath
}
\examples{
library(STexampleData)
spe <- Visium_humanDLPFC()
my.image.as.raster <- SpatialExperiment::imgRaster(spe, 
sample_id = imgData(spe)$sample_id[1], image_id = "lowres")
png("test.png")
plot(my.image.as.raster)
dev.off()
my.image.as.raster <- rasterizeFromFile("test.png")

}
