% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rasterizeFromFile}
\alias{rasterizeFromFile}
\title{Rasterize raw image from filepath}
\usage{
rasterizeFromFile(path.to.file)
}
\arguments{
\item{path.to.file}{Path to an image file.}
}
\value{
A raster image object.
}
\description{
Rasterize raw image from filepath
}
\examples{
print('rasterizeFromFile')
if (FALSE){

my.image.as.raster <- rasterizeFromFile(path.to.file)  

}
}
