% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{receptors<-,BSRInference-method}
\alias{receptors<-,BSRInference-method}
\title{receptors setter (internal use only)}
\usage{
\S4method{receptors}{BSRInference}(x) <- value
}
\description{
receptors setter (internal use only)
}
