% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R,
%   R/BSRInferenceComp-class.R
\name{reduceToBestPathway}
\alias{reduceToBestPathway}
\alias{reduceToBestPathway,BSRInference-method}
\alias{reduceToBestPathway,BSRInferenceComp-method}
\title{Keep one pathway per ligand-receptor pair}
\usage{
\S4method{reduceToBestPathway}{BSRInference}(obj)

\S4method{reduceToBestPathway}{BSRInferenceComp}(obj)
}
\arguments{
\item{obj}{BSRInferenceComp object}
}
\value{
A BSRInference object reduced to only report one pathway per
ligand-receptor pair. The pathway with the
smallest P-value is selected.

A BSRInferenceComp object reduced to only report one pathway per
ligand-receptor pair. The pathway with the
smallest P-value is selected.
}
\description{
Keep one pathway per ligand-receptor pair

Keep one pathway per ligand-receptor pair
}
\details{
Ligand-receptor pairs
are evaluated in relation with pathways that allow checking receptor
downstream correlations. It is thus possible
that several pathways are reported for a same LR pair.

Ligand-receptor pairs
are evaluated in relation with pathways that allow checking receptor
downstream correlations. It is thus possible
that several pathways are reported for a same LR pair.
}
\examples{
data(bsrinf, package = "BulkSignalR")
bsrinf.redBP <- reduceToBestPathway(bsrinf)

data(sdc, package = "BulkSignalR")
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[, -normal])

bsrdm.comp <- as(bsrdm, "BSRDataModelComp")

colA <- as.integer(1:2)
colB <- as.integer(3:4)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(
    pval = runif(n), logFC = rnorm(n, 0, 2),
    expr = runif(n, 0, 10)
)
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison

subset <- c("REACTOME_BASIGIN_INTERACTIONS",
"REACTOME_SYNDECAN_INTERACTIONS",
"REACTOME_ECM_PROTEOGLYCANS",
"REACTOME_CELL_JUNCTION_ORGANIZATION")

BulkSignalR_Reactome <- BulkSignalR_Reactome[
BulkSignalR_Reactome$`Reactome name` \%in\% subset,]

bsrinf <- initialInference(bsrdm.comp, max.pval = 1, 
"random.example", reference="REACTOME")

reduceToBestPathway(bsrinf)

}
