% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R,
%   R/BSRInferenceComp-class.R
\name{reduceToBestPathway}
\alias{reduceToBestPathway}
\alias{reduceToBestPathway,BSRInference-method}
\alias{reduceToBestPathway,BSRInferenceComp-method}
\title{Keep one pathway per ligand-receptor pair}
\usage{
\S4method{reduceToBestPathway}{BSRInference}(obj)

\S4method{reduceToBestPathway}{BSRInferenceComp}(obj)
}
\arguments{
\item{obj}{BSRInferenceComp object}
}
\value{
A BSRInference object reduced to only report one pathway per
ligand-receptor pair. The pathway with the
smallest P-value is selected.

A BSRInferenceComp object reduced to only report one pathway per
ligand-receptor pair. The pathway with the
smallest P-value is selected.
}
\description{
Keep one pathway per ligand-receptor pair

Keep one pathway per ligand-receptor pair
}
\details{
Ligand-receptor pairs
are evaluated in relation with pathways that allow checking receptor
downstream correlations. It is thus possible
that several pathways are reported for a same LR pair.

Ligand-receptor pairs
are evaluated in relation with pathways that allow checking receptor
downstream correlations. It is thus possible
that several pathways are reported for a same LR pair.
}
\examples{
print('reduceToBestPathway')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
bsrinf.redBP  <- reduceToBestPathway(bsrinf)  

print('reduceToBestPathway')

# prepare data
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:5)
colB <- as.integer(8:15)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(pval=runif(n), logFC=rnorm(n, 0, 2))
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison
bsrinf <- initialInference(bsrdm.comp,max.pval=1, "random.example")

# reduction
bsrinf.redBP  <- reduceToBestPathway(bsrinf)

}
