% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R,
%   R/BSRInferenceComp-class.R
\name{reduceToLigand}
\alias{reduceToLigand}
\alias{reduceToLigand,BSRInference-method}
\alias{reduceToLigand,BSRInferenceComp-method}
\title{Aggregate the receptors of a same ligand}
\usage{
\S4method{reduceToLigand}{BSRInference}(obj)

\S4method{reduceToLigand}{BSRInferenceComp}(obj)
}
\arguments{
\item{obj}{BSRInferenceComp object}
}
\value{
A BSRInference object but reduced to one row per ligand.
All the receptors are combined in a
semi-colon-separated list surrounded by curly brackets in the tabular
slot \code{LRinter}, and in vectors in the \code{ligands} (list) slot.

The reported P-value and target genes are those from the pathway with
the smallest P-value.

A BSRInferenceComp object but reduced to one row per ligand.
All the receptors are combined in a
semi-colon-separated list surrounded by curly brackets in the tabular
slot \code{LRinter}, and in vectors in the \code{ligands} (list) slot.

The reported P-value and target genes are those from the pathway with
the smallest P-value.
The same logic applies to the LR-score, and the receptor
expression.
}
\description{
Simplifies a ligand-receptor table to focus on the ligands.

Simplifies a ligand-receptor table to focus on the ligands.
}
\examples{
data(bsrinf, package = "BulkSignalR")

bsrinf.redL <- reduceToLigand(bsrinf)

data(sdc, package = "BulkSignalR")
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[, -normal])

bsrdm.comp <- as(bsrdm, "BSRDataModelComp")

colA <- as.integer(1:2)
colB <- as.integer(3:4)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(
    pval = runif(n), logFC = rnorm(n, 0, 2),
    expr = runif(n, 0, 10)
)
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison

subset <- c("REACTOME_BASIGIN_INTERACTIONS",
"REACTOME_SYNDECAN_INTERACTIONS",
"REACTOME_ECM_PROTEOGLYCANS",
"REACTOME_CELL_JUNCTION_ORGANIZATION")

BulkSignalR_Reactome <- BulkSignalR_Reactome[
BulkSignalR_Reactome$`Reactome name` \%in\% subset,]

bsrinf <- initialInference(bsrdm.comp, max.pval = 1, 
"random.example", reference="REACTOME")

# reduction
bsrinf.redL <- reduceToLigand(bsrinf)
}
