% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R,
%   R/BSRInferenceComp-class.R
\name{reduceToLigand}
\alias{reduceToLigand}
\alias{reduceToLigand,BSRInference-method}
\alias{reduceToLigand,BSRInferenceComp-method}
\title{Aggregate the receptors of a same ligand}
\usage{
\S4method{reduceToLigand}{BSRInference}(obj)

\S4method{reduceToLigand}{BSRInferenceComp}(obj)
}
\arguments{
\item{obj}{BSRInferenceComp object}
}
\value{
A BSRInference object but reduced to one row per ligand.
All the receptors are combined in a
semi-colon-separated list surrounded by curly brackets in the tabular
slot \code{LRinter}, and in vectors in the \code{ligands} (list) slot.

The reported P-value and target genes are those from the pathway with
the smallest P-value.

A BSRInferenceComp object but reduced to one row per ligand.
All the receptors are combined in a
semi-colon-separated list surrounded by curly brackets in the tabular
slot \code{LRinter}, and in vectors in the \code{ligands} (list) slot.

The reported P-value and target genes are those from the pathway with
the smallest P-value.
}
\description{
Simplifies a ligand-receptor table to focus on the ligands.

Simplifies a ligand-receptor table to focus on the ligands.
}
\examples{
print('reduceToLigand')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
bsrinf.redL  <- reduceToLigand(bsrinf)  

print('reduceToLigand')

# prepare data
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:5)
colB <- as.integer(8:15)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(pval=runif(n), logFC=rnorm(n, 0, 2))
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison
bsrinf <- initialInference(bsrdm.comp,max.pval=1, "random.example")

# reduction
bsrinf.redL  <- reduceToLigand(bsrinf)  

}
