% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{reduceToPathway}
\alias{reduceToPathway}
\title{Aggregate ligands and receptors at the pathway level}
\usage{
reduceToPathway(pairs)
}
\arguments{
\item{pairs}{A ligand-receptor table such as output by \code{pValuesLR} and \code{naiveBayesLR}.}
}
\value{
A table with the same structure as \code{pairs} but reduced to only report one row per pathway.
All the ligands and all the receptors forming pairs related to a certain pathway, are combined in two semi-colon-separated
lists surounded by curly braces. The information of which ligand interacted with which receptor is lost.

There is no pathway selection, i.e. potentially redundant pathways with different P-values or log-likelihood ratios (LLRs)
are all kept in the output. \code{reduceToBestPathway()} can be used to select the one with best P-value or LLR.
}
\description{
Simplifies a ligand-receptor table to focus on the pathways.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pairs.p <- pValuesLR(ds.LR,ds$param)
pairs.pw <- reduceToPathway(pairs.p)
pairs.pw.red <- reduceToBestPathway(pairs.pw)
}
}
