% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R,
%   R/BSRInferenceComp-class.R
\name{reduceToPathway}
\alias{reduceToPathway}
\alias{reduceToPathway,BSRInference-method}
\alias{reduceToPathway,BSRInferenceComp-method}
\title{Aggregate ligands and receptors at the pathway level}
\usage{
\S4method{reduceToPathway}{BSRInference}(obj)

\S4method{reduceToPathway}{BSRInferenceComp}(obj)
}
\arguments{
\item{obj}{BSRInferenceComp object}
}
\value{
A BSRInference object reduced to only report one row per pathway.
The information of which ligand interacted with which receptor is lost as
all the ligands and all the receptors forming pairs related to a certain
pathway are combined.
For a given pathway, the reported P-values and target genes are those of
the best ligand-receptor pair that
was in this pathway.
Receptors and ligands are combined in two semi-colon-separated
lists surrounded by curly brackets in the tabular slot \code{LRinter},
while the list representation slots (\code{ligands} and
\code{receptors}) are update accordingly.

A BSRInferenceComp object reduced to only report one row per pathway.
The information of which ligand interacted with which receptor is lost as
all the ligands and all the receptors forming pairs related to a certain
pathway are combined.
For a given pathway, the reported P-values and target genes are those of
the best ligand-receptor pair that
was in this pathway.
Receptors and ligands are combined in two semi-colon-separated
lists surrounded by curly brackets in the tabular slot \code{LRinter},
while the list representation slots (\code{ligands} and
\code{receptors}) are update accordingly.
}
\description{
Simplifies a ligand-receptor inference object to focus on
the pathways.

Simplifies a ligand-receptor inference object to focus on
the pathways.
}
\examples{
print('reduceToPathway')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
bsrinf.redP  <- reduceToPathway(bsrinf)  
print('reduceToPathway')

# prepare data
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:5)
colB <- as.integer(8:15)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(pval=runif(n), logFC=rnorm(n, 0, 2))
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison
bsrinf <- initialInference(bsrdm.comp,max.pval=1, "random.example")

# reduction
bsrinf.redP  <- reduceToPathway(bsrinf)  
}
