% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{reduceToReceptor}
\alias{reduceToReceptor}
\title{Aggregate the ligands of a same receptor}
\usage{
reduceToReceptor(pairs)
}
\arguments{
\item{pairs}{A ligand-receptor table such as output by \code{pValuesLR}.}
}
\value{
A table with the same structure as \code{pairs} but reduced to only report one row per receptor
within each pathway. All the ligands are combined in a semi-colon-separated list surrounded by curly braces.

The reported P-value and target genes are those from the line with the smallest P-value.
}
\description{
Simplifies a ligand-receptor table to focus on the receptors.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pairs.p <- pValuesLR(ds.LR,ds$param)
pairs.recept <- reduceToReceptor(pairs.p)
}
}
