% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R,
%   R/BSRInferenceComp-class.R
\name{reduceToReceptor}
\alias{reduceToReceptor}
\alias{reduceToReceptor,BSRInference-method}
\alias{reduceToReceptor,BSRInferenceComp-method}
\title{Aggregate the ligands of a same receptor}
\usage{
\S4method{reduceToReceptor}{BSRInference}(obj)

\S4method{reduceToReceptor}{BSRInferenceComp}(obj)
}
\arguments{
\item{obj}{BRSInferenceComp object}
}
\value{
BSRInference object reduced to one row per receptor.
All the ligands are combined in a
semi-colon-separated list surrounded by curly brackets in the tabular
slot \code{LRinter}, and in vectors in the \code{ligands} (list) slot.

The reported P-value and target genes are those from the line with the
pathway featuring the smallest P-value.

BSRInferenceComp object reduced to one row per receptor.
All the ligands are combined in a
semi-colon-separated list surrounded by curly brackets in the tabular
slot \code{LRinter}, and in vectors in the \code{ligands} (list) slot.

The reported P-value and target genes are those from the line with the
pathway featuring the smallest P-value.
}
\description{
Simplifies a ligand-receptor table to focus on the receptors.

Simplifies a ligand-receptor table to focus on the receptors.
}
\examples{
print('reduceToReceptor')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
bsrinf.redR  <- reduceToReceptor(bsrinf)  

print('reduceToReceptor')

# prepare data
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:5)
colB <- as.integer(8:15)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(pval=runif(n), logFC=rnorm(n, 0, 2))
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison
bsrinf <- initialInference(bsrdm.comp,max.pval=1, "random.example")

# reduction
bsrinf.redR  <- reduceToReceptor(bsrinf)  

}
