% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{rescoreInference}
\alias{rescoreInference}
\alias{rescoreInference,BSRInference-method}
\title{Inference re-scoring}
\usage{
\S4method{rescoreInference}{BSRInference}(
  obj,
  param,
  rank.p = 0.55,
  fdr.proc = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BY", "ABH",
    "TSBH")
)
}
\arguments{
\item{obj}{BRSInferecence object.}

\item{param}{BSRdataModel parameters.}

\item{rank.p}{A number between 0 and 1 defining the rank of the last
considered target genes.}

\item{fdr.proc}{The procedure for adjusting P-values according to
\code{\link[multtest]{mt.rawp2adjp}}.}
}
\value{
A BSRInference object.
}
\description{
A method to re-score an existing BSRInference object
(P- and Q-value estimations).
}
\details{
A BSRInference object should be created by calling 
\code{"\link[=BSRDataModel-class]{initialInference}"}

Parameters controlling the estimation
of the statistical significance of the ligand/receptor/pathway triples
are provided at the time of calling the latter method.

Nonetheless, it
might be useful to change the initially-provided parameters, in
which case this method should be called.
}
\examples{
print('rescoreInference')
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
bsrinf.new <- rescoreInference(bsrinf, param=param(bsrdm), rank.p=0.75)

}
