% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R,
%   R/BSRInferenceComp-class.R
\name{rescoreInference}
\alias{rescoreInference}
\alias{rescoreInference,BSRInference-method}
\alias{rescoreInference,BSRInferenceComp-method}
\title{Inference re-scoring}
\usage{
\S4method{rescoreInference}{BSRInference}(
  obj,
  param,
  rank.p = 0.55,
  fdr.proc = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BY", "ABH",
    "TSBH")
)

\S4method{rescoreInference}{BSRInferenceComp}(
  obj,
  param,
  rank.p = 0.55,
  fdr.proc = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BY", "ABH",
    "TSBH")
)
}
\arguments{
\item{obj}{BSRInferecenceComp object.}

\item{param}{The matching BSRDataModel parameters
(can be obtained with the accessor \code{param})}

\item{rank.p}{A number between 0 and 1 defining the rank of the last
considered target genes.}

\item{fdr.proc}{The procedure for adjusting P-values according to
\code{\link[multtest]{mt.rawp2adjp}}.}
}
\value{
A BSRInference object.

A BSRInferenceComp object.
}
\description{
A method to re-score an existing BSRInference object
(P- and Q-value estimations).

A method to re-score an existing BSRInferenceComp object
(P- and Q-value estimations).
}
\details{
A BSRInference object should be created by calling 
\code{"\link[=BSRDataModel-class]{initialInference}"}

Parameters controlling the estimation
of the statistical significance of the ligand/receptor/pathway triples
(\code{param}) are provided at the time of calling the latter method.

Nonetheless, it
might be useful to change the initially-provided parameters, in
which case this method should not be called.

A BSRInferenceComp object should be created by calling 
\code{"\link[=BSRClusterComp-class]{initialInference}"}
}
\examples{
print('rescoreInference')
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
bsrinf.new <- rescoreInference(bsrinf, param=param(bsrdm), rank.p=0.75)

print('rescoreInference')

# prepare data
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:5)
colB <- as.integer(8:15)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(pval=runif(n), logFC=rnorm(n, 0, 2))
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison
bsrinf <- initialInference(bsrdm.comp,max.pval=1, "random.example")

# rescore
bsrinf.less <- rescoreInference(bsrinf, param=param(bsrdm.comp), rank.p=0.75)

}
