% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrepare.R
\name{resetLRdb}
\alias{resetLRdb}
\title{Modify LRdb database}
\usage{
resetLRdb(db, switch = FALSE)
}
\arguments{
\item{db}{A data frame with 2 columns named
ligand and receptor.}

\item{switch}{A logical indicating whether LRdb should be extended only
(FALSE, default) or completely replaced (TRUE).}
}
\value{
Returns `NULL`, invisibly.
}
\description{
Users can provide a data frame with 2 columns named
ligand and receptor.
This can be used to extend or replace the existing
LRdb.
}
\examples{
resetLRdb(db = data.frame(ligand = "A2M", receptor = "LRP1"), switch = FALSE)
}
