% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrepare.R
\name{resetLRdb}
\alias{resetLRdb}
\title{Modify LRdb database}
\usage{
resetLRdb(db = data.frame(ligand = "A2M", receptor = "LRP1"), switch = FALSE)
}
\arguments{
\item{db}{A dataframe with 2 column names :
ligand and receptor.}

\item{switch}{By default set to FALSE, it extends the
existing LRdb database. If TRUE, LRdb is replaced by user}
}
\description{
User can define a dataframe with 2 columns named
respectively ligand and receptor.
This can be used to extand or replace the existing
LRdb.
}
\examples{
print('resetLRdb')
data(sdc,package='BulkSignalR')
resetLRdb(db=data.frame(ligand="A2M",receptor="LRP1"),switch=FALSE)
}
