% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resetNetwork}
\alias{resetNetwork}
\title{Import Network from your own}
\usage{
resetNetwork(network)
}
\arguments{
\item{network}{Network dataframe is defined with 3 columns
a.gn, b.gn & type. 'a.gn' & 'b.gn' should be gene symbols
of gene interactions. 'type'  should be set as
'controls-expression-of' when user provide
his own file.}
}
\description{
Network is a dataframe that gives relation between
genes. It's composed of 3 columns annoted as
follows :
}
\details{
a.gn : Gene Symbol 1
type : controls-expression-of
b.gn : Gene Symbol 2

When the user provide his own network
'type' should be set to 'controls-expression-of'.
}
\examples{
print("resetNetwork")
if (FALSE) {
    resetNetwork(network)
}

}
