% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInference-class.R
\name{resetToInitialOrganism}
\alias{resetToInitialOrganism}
\alias{resetToInitialOrganism,BSRInference-method}
\title{Reset gene names to initial organism providen in first instance}
\usage{
\S4method{resetToInitialOrganism}{BSRInference}(obj, conversion.dict)
}
\arguments{
\item{obj}{BSRInference object}

\item{conversion.dict}{A dictionnary}
}
\value{
An BSRInference object updated for gene names.
The gene names are replaced by the ones from
the organism providen in first instance.
}
\description{
Reset gene names to initial organism providen in first instance
}
\examples{
data(bodyMap.mouse)
ortholog.dict    <- findOrthoGenes (from_organism = "mmusculus", 
                                    from_values = rownames(bodyMap.mouse))

matrix.expression.human <- convertToHuman(counts = bodyMap.mouse, 
                                          dictionary = ortholog.dict)

bsrdm  <- prepareDataset(counts = matrix.expression.human,
          species = "mmusculus",
          conversion.dict = ortholog.dict)

bsrdm <- learnParameters(bsrdm, null.model="normal" , quick = FALSE, 
          plot.folder="./",filename = "bodyMap.mouse",
          verbose = TRUE)

bsrinf  <- initialInference(bsrdm)

bsrinf  <- resetToInitialOrganism(bsrinf, conversion.dict=ortholog.dict)

}
