% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{scalesReverse}
\alias{scalesReverse}
\title{Flip over horizontal/vertical axis}
\usage{
scalesReverse(rasterImage, axis = "x")
}
\arguments{
\item{rasterImage}{Raster image.}

\item{axis}{Axis along which the image is flipped.}
}
\value{
A flipped raster image
}
\description{
Flip over horizontal/vertical axis
}
\examples{
library(STexampleData)
spe <- Visium_humanDLPFC()
my.image.as.raster <- SpatialExperiment::imgRaster(spe, 
sample_id = imgData(spe)$sample_id[1], image_id = "lowres")
my.image.flipped <- scalesReverse(my.image.as.raster, axis = "x")
}
