% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{scalesReverse}
\alias{scalesReverse}
\title{Flip over horizontal/vertical axis}
\usage{
scalesReverse(rasterImage, axis = "x")
}
\arguments{
\item{rasterImage}{Raster image.}

\item{axis}{Axis along which the image is flipped.}
}
\value{
A flipped raster image
}
\description{
Flip over horizontal/vertical axis
}
\examples{
print('scalesReverse')
if(FALSE){
img <- png::readPNG(path.to.file)
my.image.as.raster <- grDevices::as.raster(path.to.file)  
my.image.flipped <- scalesReverse(my.image.as.raster, axis="x")

 }
}
