% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R
\name{scoreLRGeneSignatures}
\alias{scoreLRGeneSignatures}
\alias{scoreLRGeneSignatures,BSRDataModel-method}
\title{Score ligand-receptor gene signatures}
\usage{
\S4method{scoreLRGeneSignatures}{BSRDataModel}(
  obj,
  sig,
  LR.weight = 0.5,
  robust = FALSE,
  name.by.pathway = FALSE,
  abs.z.score = FALSE,
  rownames.LRP = FALSE
)
}
\arguments{
\item{obj}{A BSRDataModel object.}

\item{sig}{A BSRSignature object.}

\item{LR.weight}{A number between 0 and 1 defining the relative weight
of the ligand and the receptor in the signature.}

\item{robust}{A logical indicating that z-scores should be computed
with median and MAD instead of mean and standard deviation.}

\item{name.by.pathway}{A logical indicating whether row names of the
resulting score matrix should be complemented with (illustrative) pathway
names.}

\item{abs.z.score}{A logical to use absolute z-scores (useful if the
activity of a paythway is reported by a mixture of up- and down-genes
whose z-score averages might hide actual activity).}

\item{rownames.LRP}{Logical. If TRUE, Ligand Receptor and Pathway}
}
\value{
A matrix containing the scores of each ligand-receptor gene
signature in each sample.
}
\description{
Compute ligand-receptor gene signature scores over a BSRDataModel.
}
\examples{
print('prepareDataset')
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

print("learnParameters")
bsrdm <- learnParameters(bsrdm)
bsrdm

print('perform inference')
bsrinf <- initialInference(bsrdm)
bsrinf.redBP <- reduceToBestPathway(bsrinf) 
bsrsig.redBP <- getLRGeneSignatures(bsrinf.redBP,
qval.thres=0.001)

scoresLR <- scoreLRGeneSignatures(bsrdm,bsrsig.redBP,
                       name.by.pathway=FALSE)
}
