% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRDataModel-class.R,
%   R/BSRDataModelComp-class.R
\name{scoreLRGeneSignatures}
\alias{scoreLRGeneSignatures}
\alias{scoreLRGeneSignatures,BSRDataModel-method}
\alias{scoreLRGeneSignatures,BSRDataModelComp-method}
\title{Score ligand-receptor gene signatures}
\usage{
\S4method{scoreLRGeneSignatures}{BSRDataModel}(
  obj,
  sig,
  LR.weight = 0.5,
  robust = FALSE,
  name.by.pathway = FALSE,
  abs.z.score = FALSE,
  rownames.LRP = FALSE
)

\S4method{scoreLRGeneSignatures}{BSRDataModelComp}(
  obj,
  sig,
  LR.weight = 0.5,
  robust = FALSE,
  name.by.pathway = FALSE,
  abs.z.score = FALSE,
  rownames.LRP = FALSE
)
}
\arguments{
\item{obj}{A BSRDataModelComp object.}

\item{sig}{A BSRSignatureComp object.}

\item{LR.weight}{A number between 0 and 1 defining the relative weight
of the ligand and the receptor in the signature.}

\item{robust}{A logical indicating that z-scores should be computed
with median and MAD instead of mean and standard deviation.}

\item{name.by.pathway}{A logical indicating whether row names of the
resulting score matrix should be pathway names.}

\item{abs.z.score}{A logical to use absolute z-scores (useful if the
activity of a paythway is reported by a mixture of up- and down-genes
whose z-score averages might hide actual activity).}

\item{rownames.LRP}{A logical indicating, in case \code{name.by.pathway}
was set to TRUE, whether ligand and receptor names should be added on top.
No role if \code{name.by.pathway} was set to FALSE.}
}
\value{
A matrix containing the scores of each ligand-receptor gene
signature in each sample.

A matrix containing the scores of each ligand-receptor gene
signature in each sample.
}
\description{
Compute ligand-receptor gene signature scores over a BSRDataModel.

Compute ligand-receptor gene signature scores over a BSRDataModelComp
specific comparison.
}
\examples{
print('prepareDataset')
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

print("learnParameters")
bsrdm <- learnParameters(bsrdm)
bsrdm

print('perform inference')
bsrinf <- initialInference(bsrdm)
bsrinf.redBP <- reduceToBestPathway(bsrinf) 
bsrsig.redBP <- getLRGeneSignatures(bsrinf.redBP,
qval.thres=0.001)

scoresLR <- scoreLRGeneSignatures(bsrdm,bsrsig.redBP,
                       name.by.pathway=FALSE)
# prepare data
data(sdc,package='BulkSignalR')
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[,-normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:5)
colB <- as.integer(8:15)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(pval=runif(n), logFC=rnorm(n, 0, 2))
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison
bsrinf <- initialInference(bsrdm.comp,max.pval=1,  "random.example")

# reduction
bsrinf.red <- reduceToBestPathway(bsrinf)

# signature extraction and scoring
bsrsig.red <- getLRGeneSignatures(bsrinf.red, qval.thres=1e-6)
scores.red <- scoreLRGeneSignatures(bsrdm.comp, bsrsig.red,
                         name.by.pathway=TRUE, rownames.LRP=TRUE)

}
