% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdc}
\alias{sdc}
\title{Salivary duct carcinoma transcriptomes}
\format{
A data frame with 19764 rows and 26 variables.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE138581}
}
\usage{
data(sdc)
}
\description{
A dataset containing the read counts of salivary duct carcinomas
  and adjacent normal tissues.
}
\keyword{datasets}
