% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{signatureHeatmaps}
\alias{signatureHeatmaps}
\title{Heatmap function for gene expression of signature}
\usage{
signatureHeatmaps(
  pathway,
  bsrdm,
  bsrsig,
  h.width = 6,
  h.height = 9,
  fontsize = 6,
  show_column_names = FALSE
)
}
\arguments{
\item{pathway}{Pathway name}

\item{bsrdm}{BulkSignalR data model object.}

\item{bsrsig}{BulkSignalR signature object.
to display on screen.}

\item{h.width}{Heatmap width in cm.}

\item{h.height}{Heatmap height in cm.}

\item{fontsize}{Fontsize.}

\item{show_column_names}{Add column names on heatmap.}
}
\value{
A plot is created.

This is a convenience function
to propose a simple way
of representing expression of genes
involved in a specific pathway.
}
\description{
Generate a list of heatmaps for ligand,
receptor and target genes
for a specific pathway
}
\examples{
data(bsrdm, package = "BulkSignalR")
data(bsrinf, package = "BulkSignalR")

bsrinf.redP <- reduceToPathway(bsrinf)
bsrinf.redPBP <- reduceToBestPathway(bsrinf)
bsrsig.redPBP <- getLRGeneSignatures(bsrinf, qval.thres = 1)
pathway1 <- pathways(bsrsig.redPBP)[1]
signatureHeatmaps(
pathway = pathway1,
bsrdm = bsrdm,
bsrsig = bsrsig.redPBP,
h.width = 3,
h.height = 4,
fontsize = 1,
show_column_names = TRUE
)
}
