% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{signatureHeatmaps}
\alias{signatureHeatmaps}
\title{Heatmap function for gene expression of signature}
\usage{
signatureHeatmaps(
  pathway = "Cell surface interactions at the vascular wall",
  bsrdm,
  bsrsig,
  path = "./",
  filename = "ExpresssionSignature",
  width = 15,
  height = 8,
  format = c("svg", "png", "pdf"),
  show_column_names = FALSE
)
}
\arguments{
\item{pathway}{Pathway name}

\item{bsrdm}{BulkSignalR data model object.}

\item{bsrsig}{BulkSignalR signature object.}

\item{path}{Path to write your ouput.}

\item{filename}{An output filename.}

\item{width}{Global image width size.}

\item{height}{Global image height size.}

\item{format}{File format svg (default)/png/pdf}

\item{show_column_names}{Add column names on heatmap.}
}
\value{
A plot is created with the given filename. 

This is a convenience function 
to propose a simple way
of representing expression of genes 
involved in a specific pathway.
}
\description{
Generate a list of heatmaps for ligand,
receptor and target genes
for a specific pathway
}
\examples{
print('signatureHeatmaps')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
bsrinf.redP  <- reduceToPathway(bsrinf)  
bsrinf.redPBP <- reduceToBestPathway(bsrinf.redP) 
bsrsig.redPBP <- getLRGeneSignatures(bsrinf.redPBP,qval.thres=0.001)
pathway1 <- "Elastic fibre formation"
signatureHeatmaps(
       pathway = pathway1,
       bsrdm = bsrdm,
       bsrsig = bsrsig.redPBP,
       path = "./",
       filename = "sdc_signatureheatmap",
       width  = 15,
       height = 10 ,
       show_column_names = TRUE)
}
