% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{signatureHeatmaps}
\alias{signatureHeatmaps}
\title{Heatmap function for gene expression of signature}
\usage{
signatureHeatmaps(
  pathway,
  bsrdm,
  bsrsig,
  path = "./",
  filename = NULL,
  h.width = 6,
  h.height = 9,
  fontsize = 6,
  format = c("pdf", "svg", "png"),
  show_column_names = FALSE
)
}
\arguments{
\item{pathway}{Pathway name}

\item{bsrdm}{BulkSignalR data model object.}

\item{bsrsig}{BulkSignalR signature object.}

\item{path}{Path to write your ouput.}

\item{filename}{An output filename, NULL by default
to display on screen.}

\item{h.width}{Heatmap width in cm.}

\item{h.height}{Heatmap height in cm.}

\item{fontsize}{Fontsize.}

\item{format}{File format.}

\item{show_column_names}{Add column names on heatmap.}
}
\value{
A plot is created.

This is a convenience function 
to propose a simple way
of representing expression of genes 
involved in a specific pathway.
}
\description{
Generate a list of heatmaps for ligand,
receptor and target genes
for a specific pathway
}
\examples{
print('signatureHeatmaps')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
bsrinf.redP  <- reduceToPathway(bsrinf)  
bsrinf.redPBP <- reduceToBestPathway(bsrinf.redP) 
bsrsig.redPBP <- getLRGeneSignatures(bsrinf.redPBP,qval.thres=0.001)
pathway1 <- "Elastic fibre formation"
signatureHeatmaps(
       pathway = pathway1,
       bsrdm = bsrdm,
       bsrsig = bsrsig.redPBP,
       path = "./",
       filename = "sdc_signatureheatmap",
       h.width=6,
       h.height=9,
       fontsize=6,
       show_column_names = TRUE)
}
