% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{simpleHeatmap}
\alias{simpleHeatmap}
\title{Heatmap function for LR scores}
\usage{
simpleHeatmap(
  mat.c,
  file.name = NULL,
  dend.row = NULL,
  dend.spl = NULL,
  cols = NULL,
  width,
  height,
  pointsize = 4,
  bottom.annotation = NULL,
  n.col.clust = 0,
  n.row.clust = 0,
  gap.size = 0.5,
  cut.p = 0.01,
  row.names = TRUE
)
}
\arguments{
\item{mat.c}{A matrix with the signature scores such as output by \code{scoreLRGeneSignatures()}.}

\item{file.name}{A PDF file name.}

\item{dend.row}{A precomputed row dendrogram.}

\item{dend.spl}{A precompute sample (column) dendrogram.}

\item{cols}{A vector of colors to use for the heatmap.}

\item{width}{PDF width.}

\item{height}{PDF height.}

\item{pointsize}{PDF pointsize.}

\item{bottom.annotation}{\code{ComplexHeatmap} package bottom annotations.}

\item{n.col.clust}{Number of column clusters.}

\item{n.row.clust}{Number of row clusters.}

\item{gap.size}{Gap size between clusters.}

\item{cut.p}{Proportion of top and bottom values for thresholding.}

\item{row.names}{A logical to turn on/off the display of row (gene) names.}
}
\value{
A heatmap. Since heatmap plotting tend to be slow on the screen, it is advisable to provide a
PDF file name and plot in a file (much faster).

Extreme values (top and bottom) can be replaced by global quantiles at \code{cut.p} and \code{1-cut.p}
to avoid color scales shrunk by a few outliers.

This is a convenience function that relies on the \code{ComplexHeatmap} package to propose a simple way
of representing signature scores. If more advanced features are needed or more graphic parameters
should be controlled, users should implement their own function, e.g., starting from the code
of \code{simpleHeatmap()}.
}
\description{
Generate a PDF file with a heatmap representing ligand-receptor gene signature scores.
}
\examples{
\dontrun{
data(sdc,package="BulkSignalR")
sample.types <- rep("tumor",ncol(sdc))
sample.types[grep("^N",names(sdc),perl=TRUE)] <- "normal"
ds <- prepareDataset(sdc,sample.types)
ds <- learnParameters(ds)

# since model training takes time and it can be reused, it is advisable to save the model
save(ds,file="...",compress="bzip2")

ds.LR <- getCorrelatedLR(ds)
ds.LR <- checkReceptorSignaling(ds,ds.LR)
pairs.p <- pValuesLR(ds.LR,ds$param)
p.red.P <- reduceToPathway(pairs.p)
signatures <- getPathwayStats(p.red.P,qval.thres=0.01)
scores <- scoreLRGeneSignatures(ds,signatures,rename.by.pathway=TRUE)
simpleHeatmap(scores,"example.pdf",width=9,height=5)
}
}
