% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{simpleHeatmap}
\alias{simpleHeatmap}
\title{Heatmap function for LR scores}
\usage{
simpleHeatmap(
  mat.c,
  width,
  height,
  path = "./",
  filename = "simpleHeatmap",
  dend.row = NULL,
  dend.spl = NULL,
  cols = NULL,
  pointsize = 4,
  bottom.annotation = NULL,
  n.col.clust = 0,
  n.row.clust = 0,
  gap.size = 0.5,
  cut.p = 0.01,
  row.names = TRUE,
  column.names = TRUE,
  hcl_palette = NULL,
  reverse = FALSE,
  format = c("svg", "png", "pdf")
)
}
\arguments{
\item{mat.c}{A matrix with the signature scores such as output by
\code{scoreLRGeneSignatures()}.}

\item{width}{PDF width.}

\item{height}{PDF height.}

\item{path}{directory where to plot file.}

\item{filename}{file name.}

\item{dend.row}{A precomputed row dendrogram.}

\item{dend.spl}{A precompute sample (column) dendrogram.}

\item{cols}{A vector of colors to use for the heatmap.}

\item{pointsize}{PDF pointsize.}

\item{bottom.annotation}{\code{ComplexHeatmap} package bottom annotations.}

\item{n.col.clust}{Number of column clusters.}

\item{n.row.clust}{Number of row clusters.}

\item{gap.size}{Gap size between clusters.}

\item{cut.p}{Proportion of top and bottom values for thresholding.}

\item{row.names}{A logical to turn on/off the display}

\item{column.names}{A logical to turn on/off the display
of column (sample) names.}

\item{hcl_palette}{support for HCL colormaps in ComplexHeatmap
using color mapping function with circlize::colorRamp2().
palettes are listed in grDevides::hcl.pals().
of row (gene) names.}

\item{reverse}{A logicial to reverse or not colors in hclpalette.}

\item{format}{File format svg (default)/png/pdf}
}
\value{
A heatmap. Since heatmap plotting tend to be slow on the screen,
it is advisable to provide a
PDF file name and plot in a file (much faster).

If hcl_palette is set. colors parameter won't be used. 

Extreme values (top and bottom) can be replaced by global quantiles at
\code{cut.p} and \code{1-cut.p}
to avoid color scales shrunk by a few outliers.

This is a convenience function that relies on the \code{ComplexHeatmap}
package to propose a simple way
of representing signature scores. If more advanced features are needed or
more graphic parameters
should be controlled, users should implement their own function.
}
\description{
Generate a PDF file with a heatmap representing ligand-receptor gene
signature scores.
}
\examples{
print('simpleHeatmap')
data(sdc,package='BulkSignalR')
bsrdm <- prepareDataset(counts = sdc)
bsrdm <- learnParameters(bsrdm, 
         null.model = "normal",
         quick = FALSE, 
         plot.folder = "./",
         filename = "sdc",
         verbose = TRUE)
bsrinf <- initialInference(bsrdm)
bsrinf.redBP <- reduceToBestPathway(bsrinf) 
bsrsig.redBP <- getLRGeneSignatures(bsrinf.redBP,
qval.thres=0.001)

scoresLR <- scoreLRGeneSignatures(bsrdm,bsrsig.redBP,
                       name.by.pathway=FALSE)
simpleHeatmap(scoresLR[1:20,], 
                  path = "./",
                  filename = "sdc_scoresLR",
                  column.names = TRUE, 
                  height = 5, width = 9,
                  pointsize = 10,
                  hcl_palette = "Cividis"                   
                  )
}
