% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRClusterComp-class.R
\name{stats}
\alias{stats}
\alias{stats,BSRClusterComp-method}
\title{Cluster comparison statistics accessor}
\usage{
\S4method{stats}{BSRClusterComp}(x)
}
\arguments{
\item{x}{BSRClusterComp object}
}
\description{
Cluster comparison statistics accessor
}
