% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R,
%   R/BSRSignatureComp-class.R
\name{tgExpr}
\alias{tgExpr}
\alias{tgExpr,BSRInferenceComp-method}
\alias{tgExpr,BSRSignatureComp-method}
\title{Target gene expression accessor}
\usage{
\S4method{tgExpr}{BSRInferenceComp}(x)

\S4method{tgExpr}{BSRSignatureComp}(x)
}
\arguments{
\item{x}{BSRSignatureComp object}
}
\value{
tgExpr

tg.expr
}
\description{
Target gene expression accessor

Target gene expression accessor
}
\examples{
if(FALSE){
# prepare data
data(sdc, package = "BulkSignalR")
normal <- grep("^N", names(sdc))
bsrdm <- prepareDataset(sdc[, -normal])

# define the comparison
bsrdm.comp <- as.BSRDataModelComp(bsrdm)
colA <- as.integer(1:5)
colB <- as.integer(8:15)
n <- nrow(ncounts(bsrdm.comp))
stats <- data.frame(
    pval = runif(n), logFC = rnorm(n, 0, 2),
    expr = runif(n, 0, 10)
)
rownames(stats) <- rownames(ncounts(bsrdm.comp))
bsrcc <- defineClusterComp(bsrdm.comp, colA, colB, stats)
bsrdm.comp <- addClusterComp(bsrdm.comp, bsrcc, "random.example")

# infer ligand-receptor interactions from the comparison
bsrinf <- initialInference(bsrdm.comp, max.pval = 1, "random.example")

tgExpr(bsrinf)
}
if(FALSE){
}
}
