% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R,
%   R/BSRSignatureComp-class.R
\name{tgExpr}
\alias{tgExpr}
\alias{tgExpr,BSRInferenceComp-method}
\alias{tgExpr,BSRSignatureComp-method}
\title{Target gene expression accessor}
\usage{
\S4method{tgExpr}{BSRInferenceComp}(x)

\S4method{tgExpr}{BSRSignatureComp}(x)
}
\arguments{
\item{x}{BSRSignatureComp object}
}
\value{
tgExpr

tg.expr
}
\description{
Target gene expression accessor

Target gene expression accessor
}
\examples{
bsrinf <- new("BSRInferenceComp")
tgExpr(bsrinf)
}
