% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R,
%   R/BSRSignatureComp-class.R
\name{tgLogFC}
\alias{tgLogFC}
\alias{tgLogFC,BSRInferenceComp-method}
\alias{tgLogFC,BSRSignatureComp-method}
\title{Target gene logFC accessor}
\usage{
\S4method{tgLogFC}{BSRInferenceComp}(x)

\S4method{tgLogFC}{BSRSignatureComp}(x)
}
\arguments{
\item{x}{BSRSignatureComp object}
}
\value{
tgLogFC

tg.logFC
}
\description{
Target gene logFC accessor

Target gene logFC accessor
}
\examples{
bsrinf <- new("BSRInferenceComp")
tgLogFC(bsrinf)
}
