% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSRInferenceComp-class.R,
%   R/BSRSignatureComp-class.R
\name{tgPval}
\alias{tgPval}
\alias{tgPval,BSRInferenceComp-method}
\alias{tgPval,BSRSignatureComp-method}
\title{Target gene P-values accessor}
\usage{
\S4method{tgPval}{BSRInferenceComp}(x)

\S4method{tgPval}{BSRSignatureComp}(x)
}
\arguments{
\item{x}{BSRSignature}
}
\description{
Target gene P-values accessor

Target gene P-values accessor
}
